% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_utils.R
\name{.abort_not_tibblifiable}
\alias{.abort_not_tibblifiable}
\title{Abort when x is neither an object nor a list of objects}
\usage{
.abort_not_tibblifiable(x, arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{arg}{(\code{character(1)}) An argument name. This name will be mentioned in
error messages as the input that is at the origin of a problem.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
Nothing. Called for its side effect of throwing an error.
}
\description{
Abort when x is neither an object nor a list of objects
}
\keyword{internal}
