% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.apply_nest_lvl}
\alias{.apply_nest_lvl}
\title{Apply nesting at a single tree level}
\usage{
.apply_nest_lvl(x_split, cur_lvl, children, parent_col, id_col, children_to)
}
\arguments{
\item{x_split}{(\code{data.frame}) A data frame that has been split using
\code{\link[vctrs:vec_split]{vctrs::vec_split()}}.}

\item{cur_lvl}{(\code{integer(1)}) The key being split.}

\item{children}{(\code{data.frame}) The data frame that contains potential
children.}

\item{parent_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
identifies the parent id of each observation. Each value must either be
missing (for the root elements) or appear in the \code{id_col} column.}

\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{children_to}{(\code{character(1)}) The column name in which to store the
children.}
}
\value{
A data frame of parents with their children nested into a column.
}
\description{
Apply nesting at a single tree level
}
\keyword{internal}
