% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.assemble_ancestors_col}
\alias{.assemble_ancestors_col}
\title{Assemble the ancestors column for \code{unnest_tree()} output}
\usage{
.assemble_ancestors_col(out, ancestors_to, tree_levels, call)
}
\arguments{
\item{out}{(\code{data.frame}) The partially assembled output.}

\item{ancestors_to}{(\code{character(1)}) The column name (\code{NULL} by default) in
which to store the ids of the ancestors of a deeply nested child. Use
\code{NULL} if you don't need this information.}

\item{tree_levels}{(\code{list}) The object returned by \code{.collect_tree_levels()},
containing \code{level_data}, \code{out_ptype}, \code{level_sizes}, \code{level_parent_ids},
and \code{level_ancestors}.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A vector of ancestor ids with the same length as \code{nrow(out)}.
}
\description{
Assemble the ancestors column for \code{unnest_tree()} output
}
\keyword{internal}
