% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.assemble_parent_col}
\alias{.assemble_parent_col}
\title{Assemble the parent id column for \code{unnest_tree()} output}
\usage{
.assemble_parent_col(out, parent_to, tree_levels, id_col, call)
}
\arguments{
\item{out}{(\code{data.frame}) The partially assembled output.}

\item{parent_to}{(\code{character(1)}) The column name (\code{"parent"} by default) in
which to store the parent id of an observation. Use \code{NULL} if you don't
need this information.}

\item{tree_levels}{(\code{list}) The object returned by \code{.collect_tree_levels()},
containing \code{level_data}, \code{out_ptype}, \code{level_sizes}, \code{level_parent_ids},
and \code{level_ancestors}.}

\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A vector of parent ids with the same type as \code{out[[id_col]]} and
the same length as \code{nrow(out)}.
}
\description{
Assemble the parent id column for \code{unnest_tree()} output
}
\keyword{internal}
