% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.assemble_tree_output}
\alias{.assemble_tree_output}
\title{Assemble the final output from per-level data}
\usage{
.assemble_tree_output(
  tree_levels,
  id_col,
  level_to,
  parent_to,
  ancestors_to,
  call = caller_env()
)
}
\arguments{
\item{tree_levels}{(\code{list}) The object returned by \code{.collect_tree_levels()},
containing \code{level_data}, \code{out_ptype}, \code{level_sizes}, \code{level_parent_ids},
and \code{level_ancestors}.}

\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{level_to}{(\code{character(1)}) The column name (\code{"level"} by default) in
which to store the level of an observation. Use \code{NULL} if you don't need
this information.}

\item{parent_to}{(\code{character(1)}) The column name (\code{"parent"} by default) in
which to store the parent id of an observation. Use \code{NULL} if you don't
need this information.}

\item{ancestors_to}{(\code{character(1)}) The column name (\code{NULL} by default) in
which to store the ids of the ancestors of a deeply nested child. Use
\code{NULL} if you don't need this information.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{data.frame}) The assembled output with metadata columns appended.
}
\description{
Assemble the final output from per-level data
}
\keyword{internal}
