% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.backtick}
\alias{.backtick}
\title{Backtick-wrap a character vector}
\usage{
.backtick(x)
}
\arguments{
\item{x}{(\code{character}) A character vector to backtick-wrap.}
}
\value{
A character vector with each element wrapped in backticks and any
existing backticks escaped with a backslash.
}
\description{
Backtick-wrap a character vector
}
\keyword{internal}
