% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.check_children_to}
\alias{.check_children_to}
\title{Confirm that \code{children_to} is a single string that is different from \code{id_col} and \code{parent_col}}
\usage{
.check_children_to(children_to, id_col, parent_col, call = caller_env())
}
\arguments{
\item{children_to}{(\code{character(1)}) The column name in which to store the
children.}

\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{parent_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
identifies the parent id of each observation. Each value must either be
missing (for the root elements) or appear in the \code{id_col} column.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The input \code{children_to} object as a string.
}
\description{
Confirm that \code{children_to} is a single string that is different from \code{id_col} and \code{parent_col}
}
\keyword{internal}
