% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.check_col_new}
\alias{.check_col_new}
\title{Check that a column name does not already exist in a data frame}
\usage{
.check_col_new(
  x,
  col_name,
  col_arg = caller_arg(col_name),
  x_arg = "x",
  call = caller_env()
)
}
\arguments{
\item{x}{(\code{data.frame}) The data frame to unnest.}

\item{col_name}{(\code{character(1)}) The name of the column.}

\item{col_arg}{(\code{character(1)}) The name of the \code{col} argument, used for
error messages.}

\item{x_arg}{(\code{character(1)}) Data frame argument name used in error
messages.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{x} (invisibly). Throws an error if \code{col_name} is already a column
of \code{x}.
}
\description{
Check that a column name does not already exist in a data frame
}
\keyword{internal}
