% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.check_col_values_missing}
\alias{.check_col_values_missing}
\title{Check that a column has no missing values}
\usage{
.check_col_values_missing(x, x_arg, call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The input \code{x}.
}
\description{
Check that a column has no missing values
}
\keyword{internal}
