% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_to_tspec.R
\name{.check_field_exists}
\alias{.check_field_exists}
\title{Check that a named field exists in a spec}
\usage{
.check_field_exists(spec, name, .call = caller_env())
}
\arguments{
\item{spec}{(\code{tspec} or \code{NULL}) A spec object describing the structure of
\code{x}.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{NULL} (invisibly). Throws an error if the field does not exist.
}
\description{
Check that a named field exists in a spec
}
\keyword{internal}
