% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.check_id_col}
\alias{.check_id_col}
\title{Check that id column has no missing or duplicate values}
\usage{
.check_id_col(id_col, x, call = caller_env())
}
\arguments{
\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{x}{(\code{any}) The object to check.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The integer index of the id column.
}
\description{
Check that id column has no missing or duplicate values
}
\keyword{internal}
