% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.check_is_df}
\alias{.check_is_df}
\title{Confirm that an object is a data frame}
\usage{
.check_is_df(x, x_arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{x_arg}{(\code{character(1)}) The name of the x argument. This name will be
mentioned in error messages as the input that is at the origin of a
problem.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The input object.
}
\description{
Confirm that an object is a data frame
}
\keyword{internal}
