% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.check_key}
\alias{.check_key}
\title{Check if a key is valid}
\usage{
.check_key(.key, .call = caller_env())
}
\arguments{
\item{.key}{(\code{character}) The path of names to the field in the object.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{NULL}) Throws an error if invalid.
}
\description{
Check if a key is valid
}
\keyword{internal}
