% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untibblify.R
\name{.check_key_is_character}
\alias{.check_key_is_character}
\title{Check that a key is a character string}
\usage{
.check_key_is_character(key, call = caller_env())
}
\arguments{
\item{key}{(\code{character}) The spec key to validate.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{key} if valid; otherwise throws an error.
}
\description{
Check that a key is a character string
}
\keyword{internal}
