% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_list}
\alias{.check_list}
\title{Validate that an input is a list}
\usage{
.check_list(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{...}{Additional arguments passed to \code{\link[rlang:stop_input_type]{rlang::stop_input_type()}}.}

\item{allow_null}{(\code{logical(1)}) Whether \code{NULL} is accepted.}

\item{arg}{(\code{character(1)}) An argument name. This name will be mentioned in
error messages as the input that is at the origin of a problem.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{NULL} (invisibly) if valid; otherwise throws an error.
}
\description{
Validate that an input is a list
}
\keyword{internal}
