% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.check_lvls}
\alias{.check_lvls}
\title{Check that all elements have assigned tree levels}
\usage{
.check_lvls(lvls, call = caller_env())
}
\arguments{
\item{lvls}{(\code{integer}) A vector of levels.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The input \code{lvls}.
}
\description{
Check that all elements have assigned tree levels
}
\keyword{internal}
