% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object.R
\name{.check_named}
\alias{.check_named}
\title{Abort for missing names}
\usage{
.check_named(x, call)
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Abort for missing names
}
\keyword{internal}
