% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object.R
\name{.check_names_not_duplicated}
\alias{.check_names_not_duplicated}
\title{Abort for duplicate names}
\usage{
.check_names_not_duplicated(x, call)
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Abort for duplicate names
}
\keyword{internal}
