% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspec.R
\name{.check_names_to}
\alias{.check_names_to}
\title{Check that \code{.names_to} is valid for the given input form}
\usage{
.check_names_to(.names_to, .input_form, .call = caller_env())
}
\arguments{
\item{.names_to}{(\code{character(1)} or \code{NULL}) The name of the column in the
output which will contain the names of top-level elements of the input
named list. If \code{NULL}, the default, no name column is created.}

\item{.input_form}{(\code{character(1)}) The input form of data-frame-like lists.
Can be one of:
\itemize{
\item \code{"rowmajor"}: The default. The input is a named list of rows.
\item \code{"colmajor"}: The input is a named list of columns.
}}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Check that \code{.names_to} is valid for the given input form
}
\keyword{internal}
