% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object.R
\name{.check_not_df}
\alias{.check_not_df}
\title{Abort if \code{x} is a data frame}
\usage{
.check_not_df(x, call)
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{x} (invisibly).
}
\description{
Abort if \code{x} is a data frame
}
\keyword{internal}
