% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_utils.R
\name{.check_object_list}
\alias{.check_object_list}
\title{bort if \code{x} is not a list of objects}
\usage{
.check_object_list(x, arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{arg}{(\code{character(1)}) An argument name. This name will be mentioned in
error messages as the input that is at the origin of a problem.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{x} (invisibly). Called for side effect.
}
\description{
bort if \code{x} is not a list of objects
}
\keyword{internal}
