% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi-read_spec.R
\name{.check_openapi_version}
\alias{.check_openapi_version}
\title{Ensure the OpenAPI version is supported}
\usage{
.check_openapi_version(openapi_spec)
}
\arguments{
\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
The input \code{openapi_spec} if the version is supported.
}
\description{
Ensure the OpenAPI version is supported
}
\keyword{internal}
