% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.check_parent_id_missing}
\alias{.check_parent_id_missing}
\title{Check that parent ids aren't missing}
\usage{
.check_parent_id_missing(parent_ids, ids, call = caller_env())
}
\arguments{
\item{parent_ids}{(\code{character} or \code{integer}) The parent ids to check.}

\item{ids}{(\code{character} or \code{integer}) The potential child ids to compare
against.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The input \code{parent_ids}.
}
\description{
Check that parent ids aren't missing
}
\keyword{internal}
