% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.check_parent_ids}
\alias{.check_parent_ids}
\title{Validate parent ids: cast to id type, check for self-references and missing roots}
\usage{
.check_parent_ids(x, parent_col, id_col, id_col_name, call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{parent_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
identifies the parent id of each observation. Each value must either be
missing (for the root elements) or appear in the \code{id_col} column.}

\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{id_col_name}{(\code{character(1)}) The name of the column that uniquely
identifies each observation.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The parent ids, cast to the same type as ids.
}
\description{
Validate parent ids: cast to id type, check for self-references and missing roots
}
\keyword{internal}
