% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.check_print_names_arg}
\alias{.check_print_names_arg}
\title{Check and determine whether to print names for tibblifying}
\usage{
.check_print_names_arg(names)
}
\arguments{
\item{names}{(\code{logical(1)}) Should names be printed even if they can be
deduced from the spec?}
}
\value{
A logical value indicating whether to print names for tibblifying,
based on the provided \code{names} argument or the \code{tibblify.print_names} option
if \code{names} is \code{NULL}.
}
\description{
Check and determine whether to print names for tibblifying
}
\keyword{internal}
