% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.check_self_reference}
\alias{.check_self_reference}
\title{Check that no element is its own parent}
\usage{
.check_self_reference(parent_ids, ids, call = caller_env())
}
\arguments{
\item{parent_ids}{(\code{character} or \code{integer}) The parent ids to check.}

\item{ids}{(\code{character} or \code{integer}) The potential child ids to compare
against.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The input \code{parent_ids}.
}
\description{
Check that no element is its own parent
}
\keyword{internal}
