% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.check_tspec_combine_dots}
\alias{.check_tspec_combine_dots}
\title{Check dots arguments for tspec combine}
\usage{
.check_tspec_combine_dots(..., .call = caller_env())
}
\arguments{
\item{...}{Specifications to combine.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{list}) A list of specifications.
}
\description{
Check dots arguments for tspec combine
}
\keyword{internal}
