% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.check_tspec_combine_type}
\alias{.check_tspec_combine_type}
\title{Check spec types for compatibility}
\usage{
.check_tspec_combine_type(spec_list, .call = caller_env())
}
\arguments{
\item{spec_list}{(\code{list}) A list of specifications.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{character(1)}) The type of the specifications.
}
\description{
Check spec types for compatibility
}
\keyword{internal}
