% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.check_unnest_col_diff}
\alias{.check_unnest_col_diff}
\title{Validate an output column name and check it differs from prior names}
\usage{
.check_unnest_col_diff(
  col_name,
  x,
  ...,
  col_arg = caller_arg(col_name),
  call = caller_env()
)
}
\arguments{
\item{col_name}{(\code{character(1)}) The name of the column.}

\item{x}{(\code{data.frame}) The data frame to unnest.}

\item{...}{Already-reserved column names that \code{col_name} must differ from.}

\item{col_arg}{(\code{character(1)}) The name of the \code{col} argument, used for
error messages.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{col_name} or \code{NULL}.
}
\description{
Validate an output column name and check it differs from prior names
}
\keyword{internal}
