% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.check_unnest_col_name}
\alias{.check_unnest_col_name}
\title{Validate and normalise an output column name}
\usage{
.check_unnest_col_name(
  col_name,
  x,
  col_arg = caller_arg(col_name),
  call = caller_env()
)
}
\arguments{
\item{col_name}{(\code{character(1)}) The name of the column.}

\item{x}{(\code{data.frame}) The data frame to unnest.}

\item{col_arg}{(\code{character(1)}) The name of the \code{col} argument, used for
error messages.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{col_name} (cast to \code{character}) or \code{NULL}.
}
\description{
Validate and normalise an output column name
}
\keyword{internal}
