% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.choose_native_ptype}
\alias{.choose_native_ptype}
\title{Choose native ptype class}
\usage{
.choose_native_ptype(.ptype, .class, .fields)
}
\arguments{
\item{.ptype}{(\code{vector(0)}) A prototype of the desired output type of the
field.}

\item{.class}{(\code{character} or \code{NULL}) Additional classes for the collector.}

\item{.fields}{(\code{list}) The fields list.}
}
\value{
(\code{character} or \code{NULL}) The native ptype class name.
}
\description{
Choose native ptype class
}
\keyword{internal}
