% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_df.R
\name{.col_guess_required}
\alias{.col_guess_required}
\title{Guess whether a field is required across a list of data frames}
\usage{
.col_guess_required(col_name, df_list)
}
\arguments{
\item{col_name}{(\code{character(1)}) The column name to check.}

\item{df_list}{(\code{list}) A list of data frames.}
}
\value{
(\code{logical(1)}) \code{TRUE} if \code{col_name} is present in every data frame,
\code{FALSE} otherwise.
}
\description{
Guess whether a field is required across a list of data frames
}
\keyword{internal}
