% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_df.R
\name{.col_to_spec}
\alias{.col_to_spec}
\title{Convert a column to a tib field specification}
\usage{
.col_to_spec(col, name, empty_list_unspecified, local_env)
}
\arguments{
\item{col}{(\code{any}) A column from a data frame, which may be a vector, a list,
or a nested data frame.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{local_env}{(\code{environment}) A local environment used to track state
across recursive calls, such as whether empty lists were encountered.}
}
\value{
A tib field specification.
}
\description{
Convert a column to a tib field specification
}
\keyword{internal}
