% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_df.R
\name{.col_to_spec_df}
\alias{.col_to_spec_df}
\title{Convert a df-typed list column to a tib_df specification}
\usage{
.col_to_spec_df(
  ptype,
  col,
  name,
  list_of_col,
  empty_list_unspecified,
  local_env
)
}
\arguments{
\item{ptype}{(\code{vector(0)}) A prototype of the desired output type of the
field.}

\item{col}{(\code{any}) A column from a data frame, which may be a vector, a list,
or a nested data frame.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{list_of_col}{(\code{logical(1)}) Whether \code{col} is a \code{list_of()} column.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{local_env}{(\code{environment}) A local environment used to track state
across recursive calls, such as whether empty lists were encountered.}
}
\value{
A \code{\link[=tib_df]{tib_df()}} specification.
}
\description{
Delegates to \code{\link[=.list_of_col_to_spec_df]{.list_of_col_to_spec_df()}} or \code{\link[=.non_list_of_col_to_spec_df]{.non_list_of_col_to_spec_df()}}
based on whether \code{col} is a \code{list_of()} column.
}
\keyword{internal}
