% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.collapse_with_pad}
\alias{.collapse_with_pad}
\title{Collapse expressions with padding and optional multi-line formatting}
\usage{
.collapse_with_pad(x, multi_line, nchar_prefix = 0, width)
}
\arguments{
\item{x}{(\code{list}) Expressions to collapse.}

\item{multi_line}{(\code{logical(1)}) Should the output be formatted across
multiple lines? For example, should each element of even a short list be
displayed on its own line?}

\item{nchar_prefix}{(\code{integer(1)}) The number of characters that are "used
up" by any prefix portions of the output. Used to determine whether the
output will fit on a single line or if it should be wrapped across multiple
lines.}

\item{width}{(\code{integer(1)}) The width (in number of characters) of text
output to generate.}
}
\value{
A single string with the expressions collapsed, either on a single
line or wrapped across multiple lines with indentation.
}
\description{
Collapse expressions with padding and optional multi-line formatting
}
\keyword{internal}
