% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.collect_tree_levels}
\alias{.collect_tree_levels}
\title{Traverse the tree and collect per-level data}
\usage{
.collect_tree_levels(
  x,
  id_col,
  child_col,
  parent_to,
  ancestors_to,
  call = caller_env()
)
}
\arguments{
\item{x}{(\code{data.frame}) The root-level data frame.}

\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{child_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
contains the children of an observation. This column must be a list where
each element is either \code{NULL} or a data frame with the same columns as \code{x}.}

\item{parent_to}{(\code{character(1)}) The column name (\code{"parent"} by default) in
which to store the parent id of an observation. Use \code{NULL} if you don't
need this information.}

\item{ancestors_to}{(\code{character(1)}) The column name (\code{NULL} by default) in
which to store the ids of the ancestors of a deeply nested child. Use
\code{NULL} if you don't need this information.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A named list with elements \code{level_data}, \code{out_ptype}, \code{level_sizes},
\code{level_parent_ids}, and \code{level_ancestors}.
}
\description{
Traverse the tree and collect per-level data
}
\keyword{internal}
