% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.combine_processed_fields}
\alias{.combine_processed_fields}
\title{Combine processed simple and complex fields}
\usage{
.combine_processed_fields(
  spec_simple,
  spec_complex,
  coll_locations,
  is_sub,
  first_keys
)
}
\arguments{
\item{spec_simple}{(\code{list}) Processed simple fields.}

\item{spec_complex}{(\code{list}) Processed complex fields.}

\item{coll_locations}{(\code{integer}) A numeric vector of locations.}

\item{is_sub}{(\code{logical}) Which fields are sub-fields?}

\item{first_keys}{(\code{character}) First keys of the complex fields.}
}
\value{
A list with fields, locations, and keys.
}
\description{
Combine processed simple and complex fields
}
\keyword{internal}
