% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compat_map_chr}
\alias{.compat_map_chr}
\title{Map to a character vector}
\usage{
.compat_map_chr(x, .f, ...)
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{.f}{(\code{function}) Function to apply to each element of \code{x}.}

\item{...}{Additional arguments passed to \code{.f}.}
}
\value{
(\code{character}) The mapped character vector.
}
\description{
Map to a character vector
}
\keyword{internal}
