% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.data_field_name}
\alias{.data_field_name}
\title{Format a data frame column reference for error messages}
\usage{
.data_field_name(col_name)
}
\value{
A string in the form "x$col".
}
\description{
Format a data frame column reference for error messages
}
\keyword{internal}
