% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.deprecate_arg}
\alias{.deprecate_arg}
\title{Handle deprecated arguments}
\usage{
.deprecate_arg(
  good_arg,
  bad_arg,
  fn_name = rlang::call_name(rlang::caller_call()),
  pkg_version = "0.4.0",
  good_arg_name = rlang::caller_arg(good_arg),
  bad_arg_name = rlang::caller_arg(bad_arg),
  .call = rlang::caller_env(),
  user_env = rlang::caller_env(2)
)
}
\arguments{
\item{good_arg}{(\code{any}) The value of the new argument.}

\item{bad_arg}{(\code{any}) The value of the deprecated argument.}

\item{fn_name}{(\code{character(1)}) Name of the calling function.}

\item{pkg_version}{(\code{character(1)}) Package version when deprecation
occurred.}

\item{good_arg_name}{(\code{character(1)}) Name of the new argument.}

\item{bad_arg_name}{(\code{character(1)}) Name of the deprecated argument.}

\item{.call}{(\code{environment}) The environment to use for error messages.}

\item{user_env}{(\code{environment}) The environment from which the calling
function is being called.}
}
\value{
(\code{any}) The value to use.
}
\description{
Handle deprecated arguments
}
\keyword{internal}
