% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_df.R
\name{.df_guess_required}
\alias{.df_guess_required}
\title{Guess whether each field is required in a df-typed list column}
\usage{
.df_guess_required(fields_spec, col, ptype)
}
\arguments{
\item{fields_spec}{(\code{list}) A named list of tib field specifications.}

\item{col}{(\code{list}) A list column whose elements are data frames.}
}
\value{
\code{fields_spec} with \verb{$required} updated for each field.
}
\description{
Guess whether each field is required in a df-typed list column
}
\keyword{internal}
