% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.double_quote}
\alias{.double_quote}
\title{Wrap a string in double quotes if it's not \code{NULL}}
\usage{
.double_quote(x)
}
\arguments{
\item{x}{(\code{character(1)} or \code{NULL}) A string to wrap in double quotes, or
\code{NULL}.}
}
\value{
The input string wrapped in double quotes if it was not \code{NULL}, or
\code{NULL}.
}
\description{
Wrap a string in double quotes if it's not \code{NULL}
}
\keyword{internal}
