% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_utils.R
\name{.drop_empty_lists}
\alias{.drop_empty_lists}
\title{Remove empty lists from an object}
\usage{
.drop_empty_lists(x)
}
\arguments{
\item{x}{(\code{list}) An object which might contain empty lists.}
}
\value{
The input object with empty lists removed. If any were removed, the
returned object has an attribute \code{had_empty_lists} set to \code{TRUE}.
}
\description{
Remove empty lists from an object
}
\keyword{internal}
