% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.element_subset_string}
\alias{.element_subset_string}
\title{Convert a path element to a subset string}
\usage{
.element_subset_string(elt)
}
\arguments{
\item{elt}{(\code{character} or \code{integer}) An element of a path.}
}
\value{
A string representing a subset operation for the given path element,
such as \code{"$a"} or \code{"[[1]]"}.
}
\description{
Convert a path element to a subset string
}
\keyword{internal}
