% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.eval_pull}
\alias{.eval_pull}
\title{Evaluate and extract a single column selection}
\usage{
.eval_pull(x, col, col_arg, call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{col}{(\code{character}, \code{integer}, or \code{symbol}) Defused R code describing a
selection according to the tidyselect syntax.}

\item{col_arg}{(\code{character(1)}) The name of the \code{col} argument, used for
error messages.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
An integer index of the selected column, named by the column name.
}
\description{
Evaluate and extract a single column selection
}
\keyword{internal}
