% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.extract_children}
\alias{.extract_children}
\title{Extract and validate the children list from a level data frame}
\usage{
.extract_children(x, child_col, call = caller_env())
}
\arguments{
\item{x}{(\code{data.frame}) Current level data frame.}

\item{child_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
contains the children of an observation. This column must be a list where
each element is either \code{NULL} or a data frame with the same columns as \code{x}.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{list}) The children list extracted from \code{x[[child_col]]}.
}
\description{
Extract and validate the children list from a level data frame
}
\keyword{internal}
