% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.fields_from_schema_properties}
\alias{.fields_from_schema_properties}
\title{Extract tib fields from schema properties}
\usage{
.fields_from_schema_properties(schema, openapi_spec)
}
\arguments{
\item{schema}{(\code{list}) A JSON schema object, as defined in the \href{https://spec.openapis.org/oas/v3.1.0#schema-object}{Schema Object}, typically from
\code{openapi_spec$components$schemas} or inline within the spec.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A named list of tib field specs corresponding to the schema
properties, with required flags applied.
}
\description{
Extract tib fields from schema properties
}
\keyword{internal}
