% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspec.R
\name{.flatten_fields}
\alias{.flatten_fields}
\title{Flatten nested field specifications}
\usage{
.flatten_fields(.fields)
}
\arguments{
\item{.fields}{(\code{list}) A list of field specifications, typically created by
\verb{tib_*()}.}
}
\value{
A flattened, named list of field specifications.
}
\description{
Flatten nested field specifications
}
\keyword{internal}
