% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi-read_spec.R
\name{.forget_parse_schema_memoised}
\alias{.forget_parse_schema_memoised}
\title{Ensure the OpenAPI version is supported}
\usage{
.forget_parse_schema_memoised()
}
\value{
If memoise is installed, \code{TRUE} if the function is memoised, \code{FALSE}
otherwise. If memoise is not installed, \code{NULL} (invisibly).
}
\description{
Ensure the OpenAPI version is supported
}
\keyword{internal}
