% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.format_fields}
\alias{.format_fields}
\title{Format fields to print a function call}
\usage{
.format_fields(
  f_name,
  fields,
  width,
  force_names,
  args = NULL,
  fully_qualify = FALSE
)
}
\arguments{
\item{f_name}{(\code{character(1)}) The (possibly ANSI-colored) function name.}

\item{fields}{(\code{list} or \code{NULL}) Fields to format.}

\item{width}{(\code{integer(1)}) The width (in number of characters) of text
output to generate.}

\item{force_names}{(\code{logical(1)}) Should names be printed even if they can be
deduced from the spec?}

\item{args}{(\code{list} or \code{NULL}) Additional arguments to format and display
before fields.}

\item{fully_qualify}{(\code{logical(1)}) Should \verb{tib_*()} and \verb{tspec_*()} calls be
prefixed with \verb{tibblify::}?}
}
\value{
A single string with the formatted fields.
}
\description{
Format fields to print a function call
}
\keyword{internal}
