% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-tib.R
\name{.format_fill_arg}
\alias{.format_fill_arg}
\title{Format the \code{.fill} argument for display}
\usage{
.format_fill_arg(x, .fill)
}
\arguments{
\item{x}{A tibblify collector object.}

\item{.fill}{(\code{vector} or \code{NULL}) Optionally, a value to use if the field
is empty or does not exist. Note: this value must match the \code{.ptype_inner}
of the field (the value \emph{before} any transformation), not the \code{.ptype}.}
}
\value{
(\code{character(1)} or \code{NULL}) The formatted fill string, or \code{NULL} if
the argument should be omitted.
}
\description{
Format the \code{.fill} argument for display
}
\keyword{internal}
