% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.format_maybe_fully_qualify}
\alias{.format_maybe_fully_qualify}
\title{Format a function name with optional full qualification}
\usage{
.format_maybe_fully_qualify(f_name, fully_qualify)
}
\arguments{
\item{f_name}{(\code{character(1)}) The (possibly ANSI-colored) function name.}

\item{fully_qualify}{(\code{logical(1)}) Should \verb{tib_*()} and \verb{tspec_*()} calls be
prefixed with \verb{tibblify::}?}
}
\value{
The formatted function name, potentially fully qualified.
}
\description{
Format a function name with optional full qualification
}
\keyword{internal}
