% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-tib.R
\name{.format_ptype_arg}
\alias{.format_ptype_arg}
\title{Format the \code{.ptype} argument for display}
\usage{
.format_ptype_arg(x)
}
\arguments{
\item{x}{A tibblify collector object.}
}
\value{
(\code{character(1)} or \code{NULL}) The formatted ptype string, or \code{NULL} if
the argument should be omitted.
}
\description{
Format the \code{.ptype} argument for display
}
\keyword{internal}
