% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-tib.R
\name{.format_ptype_inner}
\alias{.format_ptype_inner}
\title{Format the \code{.ptype_inner} argument for display}
\usage{
.format_ptype_inner(x, .ptype_inner)
}
\arguments{
\item{x}{A tibblify collector object.}

\item{.ptype_inner}{(\code{vector(0)}) A prototype of the input field.}
}
\value{
(\code{character(1)} or \code{NULL}) The formatted \code{.ptype} string, or \code{NULL}
if the argument should be omitted.
}
\description{
Format the \code{.ptype_inner} argument for display
}
\keyword{internal}
