% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-tib.R
\name{.format_tib_f}
\alias{.format_tib_f}
\alias{.format_tib_f.tib_unspecified}
\alias{.format_tib_f.tib_scalar_logical}
\alias{.format_tib_f.tib_scalar_integer}
\alias{.format_tib_f.tib_scalar_numeric}
\alias{.format_tib_f.tib_scalar_character}
\alias{.format_tib_f.tib_scalar_date}
\alias{.format_tib_f.tib_scalar_chr_date}
\alias{.format_tib_f.tib_scalar}
\alias{.format_tib_f.tib_vector_logical}
\alias{.format_tib_f.tib_vector_integer}
\alias{.format_tib_f.tib_vector_numeric}
\alias{.format_tib_f.tib_vector_character}
\alias{.format_tib_f.tib_vector_date}
\alias{.format_tib_f.tib_vector_chr_date}
\alias{.format_tib_f.tib_vector}
\alias{.format_tib_f.tib_variant}
\alias{.format_tib_f.tib_row}
\alias{.format_tib_f.tib_df}
\alias{.format_tib_f.tib_recursive}
\alias{.format_tib_f.default}
\title{Return the display function name for a collector}
\usage{
.format_tib_f(x)

\method{.format_tib_f}{tib_unspecified}(x)

\method{.format_tib_f}{tib_scalar_logical}(x)

\method{.format_tib_f}{tib_scalar_integer}(x)

\method{.format_tib_f}{tib_scalar_numeric}(x)

\method{.format_tib_f}{tib_scalar_character}(x)

\method{.format_tib_f}{tib_scalar_date}(x)

\method{.format_tib_f}{tib_scalar_chr_date}(x)

\method{.format_tib_f}{tib_scalar}(x)

\method{.format_tib_f}{tib_vector_logical}(x)

\method{.format_tib_f}{tib_vector_integer}(x)

\method{.format_tib_f}{tib_vector_numeric}(x)

\method{.format_tib_f}{tib_vector_character}(x)

\method{.format_tib_f}{tib_vector_date}(x)

\method{.format_tib_f}{tib_vector_chr_date}(x)

\method{.format_tib_f}{tib_vector}(x)

\method{.format_tib_f}{tib_variant}(x)

\method{.format_tib_f}{tib_row}(x)

\method{.format_tib_f}{tib_df}(x)

\method{.format_tib_f}{tib_recursive}(x)

\method{.format_tib_f}{default}(x)
}
\arguments{
\item{x}{A tibblify collector object.}
}
\value{
(\code{character(1)}) The (possibly ANSI-colored) function name string.
}
\description{
Return the display function name for a collector
}
\keyword{internal}
