% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_inform_unspecified.R
\name{.format_unspecified_paths}
\alias{.format_unspecified_paths}
\title{Prep unspecified paths for messaging}
\usage{
.format_unspecified_paths(path_list, path = character())
}
\arguments{
\item{path_list}{(\code{list}) A list of \code{\link[=tib_unspecified]{tib_unspecified()}} objects and objects
containing \code{\link[=tib_unspecified]{tib_unspecified()}} objects.}

\item{path}{(\code{character}) Current path prefix.}
}
\value{
The formatted paths as \code{character}.
}
\description{
Prep unspecified paths for messaging
}
\keyword{internal}
