% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_to_tspec.R
\name{.get_nested_field}
\alias{.get_nested_field}
\title{Extract a nested (df, row, or recursive) field from a spec}
\usage{
.get_nested_field(spec, name, .call = caller_env())
}
\arguments{
\item{spec}{(\code{tspec} or \code{NULL}) A spec object describing the structure of
\code{x}.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{tib_collector}) The field spec.
}
\description{
Extract a nested (df, row, or recursive) field from a spec
}
\keyword{internal}
