% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.get_openapi_type}
\alias{.get_openapi_type}
\title{Determine the type of an OpenAPI schema object}
\usage{
.get_openapi_type(schema)
}
\arguments{
\item{schema}{(\code{list}) A JSON schema object, as defined in the
\href{https://spec.openapis.org/oas/v3.1.0#schema-object}{Schema Object},
typically from \code{openapi_spec$components$schemas} or inline within the
spec.}
}
\value{
(\code{character(1)}) The type string, one of \code{"object"}, \code{"array"},
\code{"string"}, \code{"integer"}, \code{"boolean"}, \code{"number"}, or \code{"variant"}.
}
\description{
Determine the type of an OpenAPI schema object
}
\keyword{internal}
