% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_utils.R
\name{.get_ptype_common}
\alias{.get_ptype_common}
\title{Find the common ptype of a list of objects}
\usage{
.get_ptype_common(x, empty_list_unspecified)
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}
}
\value{
A list with component \code{has_common_ptype} (\code{TRUE} if so, \code{FALSE}
otherwise) and optional components \code{ptype} (an object representing the
common \code{ptype}, if there is one) and \code{had_empty_lists} (\code{TRUE} if
\code{empty_list_unspecified} is \code{TRUE} and the \code{x} input had such empty lists).
}
\description{
Find the common ptype of a list of objects
}
\keyword{internal}
