% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object_utils.R
\name{.get_required}
\alias{.get_required}
\title{Determine which fields are required in an object list}
\usage{
.get_required(x, sample_size = 10000)
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{sample_size}{(\code{integer(1)}) Maximum number of records to sample when
\code{x} is large.}
}
\value{
A named logical vector indicating which fields are present in every
element of \code{x}.
}
\description{
Determine which fields are required in an object list
}
\keyword{internal}
