% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object.R
\name{.guess_object_field_spec_df}
\alias{.guess_object_field_spec_df}
\title{Guess the field spec for a data-frame-typed field}
\usage{
.guess_object_field_spec_df(value, name, empty_list_unspecified, local_env)
}
\arguments{
\item{value}{(\code{list}) An object list whose fields will be guessed.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{local_env}{(\code{environment}) A local environment used to track state
across recursive calls, such as whether empty lists were encountered.}
}
\value{
A \code{tib_df} field specification.
}
\description{
Guess the field spec for a data-frame-typed field
}
\keyword{internal}
