% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object_utils.R
\name{.guess_object_field_spec_object_list}
\alias{.guess_object_field_spec_object_list}
\title{Guess the spec for an object list field}
\usage{
.guess_object_field_spec_object_list(
  value,
  name,
  empty_list_unspecified,
  simplify_list,
  local_env
)
}
\arguments{
\item{value}{(\code{list}) An object list whose fields will be guessed.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{simplify_list}{(\code{logical(1)}) Should scalar lists be simplified to
vectors?}

\item{local_env}{(\code{environment}) A local environment used to track state
across recursive calls, such as whether empty lists were encountered.}
}
\value{
A \code{\link[=tib_df]{tib_df()}} spec keyed by \code{name}.
}
\description{
Guess the spec for an object list field
}
\keyword{internal}
