% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object.R
\name{.guess_object_field_spec_vector}
\alias{.guess_object_field_spec_vector}
\title{Guess the field spec for a vector-typed field}
\usage{
.guess_object_field_spec_vector(value, name)
}
\arguments{
\item{value}{(\code{list}) An object list whose fields will be guessed.}

\item{name}{(\code{character(1)}) The name of the field.}
}
\value{
A tib field specification.
}
\description{
Guess the field spec for a vector-typed field
}
\keyword{internal}
