% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object_utils.R
\name{.guess_object_list_field_spec_dont_simplify}
\alias{.guess_object_list_field_spec_dont_simplify}
\title{Guess the spec for a list field without list simplification}
\usage{
.guess_object_list_field_spec_dont_simplify(
  value,
  name,
  empty_list_unspecified,
  simplify_list,
  local_env
)
}
\arguments{
\item{value}{(\code{list}) An object list whose fields will be guessed.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{simplify_list}{(\code{logical(1)}) Should scalar lists be simplified to
vectors?}

\item{local_env}{(\code{environment}) A local environment used to track state
across recursive calls, such as whether empty lists were encountered.}
}
\value{
A tib field spec.
}
\description{
Guess the spec for a list field without list simplification
}
\keyword{internal}
