% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object_utils.R
\name{.guess_object_list_field_spec_flat_to_vector}
\alias{.guess_object_list_field_spec_flat_to_vector}
\title{Guess a vector spec from a flat list of field values}
\usage{
.guess_object_list_field_spec_flat_to_vector(value_flat, name, ptype_result)
}
\arguments{
\item{value_flat}{(\code{list}) The flattened values of a list field.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{ptype_result}{(\code{list}) The result of \code{\link[=.get_ptype_common]{.get_ptype_common()}} applied to
\code{value_flat}.}
}
\value{
A \code{\link[=tib_vector]{tib_vector()}} spec.
}
\description{
Guess a vector spec from a flat list of field values
}
\keyword{internal}
