% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object_utils.R
\name{.guess_object_list_spec}
\alias{.guess_object_list_spec}
\title{Guess field specs for an object list}
\usage{
.guess_object_list_spec(
  object_list,
  empty_list_unspecified,
  simplify_list,
  local_env
)
}
\arguments{
\item{object_list}{(\code{list}) A list of named lists (objects) whose fields
are to be guessed.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{simplify_list}{(\code{logical(1)}) Should scalar lists be simplified to
vectors?}

\item{local_env}{(\code{environment}) A local environment used to track state
across recursive calls, such as whether empty lists were encountered.}
}
\value{
A named list of tib field specs.
}
\description{
Guess field specs for an object list
}
\keyword{internal}
