% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object.R
\name{.guess_vector_input_form}
\alias{.guess_vector_input_form}
\title{Guess whether a list field can be simplified to a vector spec}
\usage{
.guess_vector_input_form(value, name)
}
\arguments{
\item{value}{(\code{list}) An object list whose fields will be guessed.}

\item{name}{(\code{character(1)}) The name of the field.}
}
\value{
A list with:
\itemize{
\item \code{can_simplify} (\code{logical(1)}): Whether the field can be simplified.
\item \code{tib_spec}: A tib field specification (present only when \code{can_simplify}
is \code{TRUE}).
}
}
\description{
Guess whether a list field can be simplified to a vector spec
}
\keyword{internal}
