% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object.R
\name{.guess_vector_input_form_field_scalar}
\alias{.guess_vector_input_form_field_scalar}
\title{Build a tib spec for a field-scalar input form}
\usage{
.guess_vector_input_form_field_scalar(value, name, ptype)
}
\arguments{
\item{value}{(\code{list}) An object list whose fields will be guessed.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{ptype}{(\code{vector(0)}) A prototype of the desired output type of the
field.}
}
\value{
A list with \code{can_simplify = TRUE} and \code{tib_spec}, a \code{tib_vector}
field specification.
}
\description{
Build a tib spec for a field-scalar input form
}
\keyword{internal}
