% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object.R
\name{.imap_guess_object_field_spec}
\alias{.imap_guess_object_field_spec}
\title{Map \code{.guess_object_field_spec} over a named list}
\usage{
.imap_guess_object_field_spec(
  x,
  empty_list_unspecified,
  simplify_list,
  local_env
)
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{simplify_list}{(\code{logical(1)}) Should scalar lists be simplified to
vectors?}

\item{local_env}{(\code{environment}) A local environment used to track state
across recursive calls, such as whether empty lists were encountered.}
}
\value{
A named list of tib field specifications, one per element of \code{x}.
}
\description{
Map \code{.guess_object_field_spec} over a named list
}
\keyword{internal}
