% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object_utils.R
\name{.is_field_scalar}
\alias{.is_field_scalar}
\title{Is every element of a field scalar?}
\usage{
.is_field_scalar(value)
}
\arguments{
\item{value}{(\code{list}) An object list whose fields will be guessed.}
}
\value{
\code{TRUE} if every element has size 1 or is \code{NULL}, \code{FALSE} otherwise.
}
\description{
Is every element of a field scalar?
}
\keyword{internal}
