% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_utils.R
\name{.is_list_of_null}
\alias{.is_list_of_null}
\title{Is x a list of NULLs?}
\usage{
.is_list_of_null(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if every element of \code{x} is \code{NULL}, \code{FALSE}
otherwise.
}
\description{
Is x a list of NULLs?
}
\keyword{internal}
