% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_utils.R
\name{.is_list_of_object_lists}
\alias{.is_list_of_object_lists}
\title{Is x a list of object lists?}
\usage{
.is_list_of_object_lists(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if every non-\code{NULL} element of \code{x} is a list
of objects, \code{FALSE} otherwise.
}
\description{
Is x a list of object lists?
}
\keyword{internal}
