% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_utils.R
\name{.is_object}
\alias{.is_object}
\title{Is x an object?}
\usage{
.is_object(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if \code{x} is an object (a fully named list with
unique names), \code{FALSE} otherwise.
}
\description{
Is x an object?
}
\keyword{internal}
