% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_utils.R
\name{.is_object_list}
\alias{.is_object_list}
\title{Is x a list of objects?}
\usage{
.is_object_list(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if \code{x} is a list of objects, \code{FALSE}
otherwise.
}
\description{
Is x a list of objects?
}
\keyword{internal}
