% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.is_tib}
\alias{.is_tib}
\title{Check if object is a tib collector}
\usage{
.is_tib(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if \code{x} is a \code{tib_collector}.
}
\description{
Check if object is a tib collector
}
\keyword{internal}
