% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.is_tib_name_canonical}
\alias{.is_tib_name_canonical}
\title{Check if a tib field has a canonical name}
\usage{
.is_tib_name_canonical(field, name)
}
\arguments{
\item{field}{(\code{list}) A tib field.}

\item{name}{(\code{character(1)}) The name to check against the field's key.}
}
\value{
\code{TRUE} if the field's key is a single string that matches \code{name},
\code{FALSE} otherwise.
}
\description{
Check if a tib field has a canonical name
}
\keyword{internal}
