% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_other.R
\name{.is_tib_row}
\alias{.is_tib_row}
\title{Check if object is a tib row}
\usage{
.is_tib_row(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if \code{x} is a \code{tib_row}.
}
\description{
Check if object is a tib row
}
\keyword{internal}
