% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.is_tib_scalar}
\alias{.is_tib_scalar}
\title{Check if object is a tib scalar}
\usage{
.is_tib_scalar(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if \code{x} is a \code{tib_scalar}.
}
\description{
Check if object is a tib scalar
}
\keyword{internal}
