% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.is_tib_unspecified}
\alias{.is_tib_unspecified}
\title{Check if object is a tib unspecified}
\usage{
.is_tib_unspecified(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if \code{x} is a \code{tib_unspecified}.
}
\description{
Check if object is a tib unspecified
}
\keyword{internal}
