% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_other.R
\name{.is_tib_variant}
\alias{.is_tib_variant}
\title{Check if object is a tib variant}
\usage{
.is_tib_variant(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if \code{x} is a \code{tib_variant}.
}
\description{
Check if object is a tib variant
}
\keyword{internal}
