% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.is_tib_vector}
\alias{.is_tib_vector}
\title{Check if object is a tib vector}
\usage{
.is_tib_vector(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if \code{x} is a \code{tib_vector}.
}
\description{
Check if object is a tib vector
}
\keyword{internal}
