% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspec.R
\name{.is_tspec}
\alias{.is_tspec}
\title{Check if object is a tibblify specification}
\usage{
.is_tspec(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if \code{x} is a \code{tspec} object.
}
\description{
Check if object is a tibblify specification
}
\keyword{internal}
