% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_utils.R
\name{.is_unspecified}
\alias{.is_unspecified}
\title{Is the object unspecified?}
\usage{
.is_unspecified(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
\code{TRUE} if the object has class \code{"vctrs_unspecified"}, \code{FALSE}
otherwise.
}
\description{
Is the object unspecified?
}
\keyword{internal}
