% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_url_string}
\alias{.is_url_string}
\title{Check whether input is an http(s) URL string}
\usage{
.is_url_string(x, arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{arg}{(\code{character(1)}) An argument name. This name will be mentioned in
error messages as the input that is at the origin of a problem.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{logical(1)}) \code{TRUE} for scalar strings starting with \verb{http://} or
\verb{https://}, \code{FALSE} otherwise.
}
\description{
Check whether input is an http(s) URL string
}
\keyword{internal}
