% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_utils.R
\name{.is_vec}
\alias{.is_vec}
\title{Is the object a vector?}
\usage{
.is_vec(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
\code{TRUE} if the object is a non-list vector, \code{FALSE} otherwise.
}
\description{
Is the object a vector?
}
\keyword{internal}
