% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_utils.R
\name{.lgl_to_bullet}
\alias{.lgl_to_bullet}
\title{Convert a logical vector to cli bullet symbols}
\usage{
.lgl_to_bullet(x)
}
\arguments{
\item{x}{(\code{logical}) A logical vector where \code{TRUE} maps to a check mark
bullet and \code{FALSE} to a cross bullet.}
}
\value{
(\code{character}) A character vector of cli bullet names (\code{"v"} or
\code{"x"}) the same length as \code{x}.
}
\description{
Convert a logical vector to cli bullet symbols
}
\keyword{internal}
