% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.list_drop_null}
\alias{.list_drop_null}
\title{Drop NULL elements from a list}
\usage{
.list_drop_null(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
\code{x} with \code{NULL} elements removed.
}
\description{
Drop NULL elements from a list
}
\keyword{internal}
