% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_utils.R
\name{.list_is_list_of_null}
\alias{.list_is_list_of_null}
\title{For each element, is it a list of NULLs?}
\usage{
.list_is_list_of_null(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical}) A logical vector the same length as \code{x}, where each
element is \code{TRUE} if the corresponding element of \code{x} is itself a list of
\code{NULL}s.
}
\description{
For each element, is it a list of NULLs?
}
\keyword{internal}
