% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_df.R
\name{.list_of_col_to_spec_df}
\alias{.list_of_col_to_spec_df}
\title{Build field specs from a list_of df column}
\usage{
.list_of_col_to_spec_df(col, ptype, empty_list_unspecified, local_env)
}
\arguments{
\item{col}{(\code{any}) A column from a data frame, which may be a vector, a list,
or a nested data frame.}

\item{ptype}{(\code{vector(0)}) A prototype of the desired output type of the
field.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{local_env}{(\code{environment}) A local environment used to track state
across recursive calls, such as whether empty lists were encountered.}
}
\value{
A named list of tib field specifications.
}
\description{
Build field specs from a list_of df column
}
\keyword{internal}
