% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.loc_name_helper}
\alias{.loc_name_helper}
\title{Format location names and types}
\usage{
.loc_name_helper(locs, types)
}
\arguments{
\item{locs}{(\code{integer}) Locations of the elements.}

\item{types}{(\code{character}) Types of those elements.}
}
\value{
(\code{character}) Dot-prefixed location names and types.
}
\description{
Format location names and types
}
\keyword{internal}
