% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_inform_unspecified.R
\name{.maybe_inform_unspecified}
\alias{.maybe_inform_unspecified}
\title{Potentially inform users about unspecified fields}
\usage{
.maybe_inform_unspecified(spec, inform_unspecified, call = caller_env())
}
\arguments{
\item{spec}{(\code{tspec}) A specification of how to convert \code{x}. Generated with
\code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}}, \code{\link[=tspec_object]{tspec_object()}}, \code{\link[=tspec_recursive]{tspec_recursive()}}, or
\code{\link[=guess_tspec]{guess_tspec()}}. If \code{spec} is \code{NULL} (the default), \code{guess_tspec(x, inform_unspecified = TRUE)} will be used to guess the \code{spec}.}

\item{inform_unspecified}{(\code{logical(1)}) Inform about fields whose type could
not be determined?}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The \code{spec} object.
}
\description{
Potentially inform users about unspecified fields
}
\keyword{internal}
