% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.name_exprs}
\alias{.name_exprs}
\title{Collapse expressions with padding and optional multi-line formatting}
\usage{
.name_exprs(exprs, names, show_name)
}
\arguments{
\item{exprs}{(\code{list}) Expressions to collapse.}

\item{names}{(\code{character}) Names corresponding to the expressions.}

\item{show_name}{(\code{logical}) A vector indicating whether each name should be
shown.}
}
\value{
A character vector of the expressions with names prepended where
\code{show_name} is \code{TRUE}. Non-syntactic names are backticked. If \code{show_name}
is \code{FALSE}, the expressions are returned without names.
}
\description{
Collapse expressions with padding and optional multi-line formatting
}
\keyword{internal}
