% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.nchar_field_names}
\alias{.nchar_field_names}
\title{Calculate the number of characters in a field's name part}
\usage{
.nchar_field_names(fields)
}
\arguments{
\item{fields}{(\code{list} or \code{NULL}) Fields to format.}
}
\value{
An integer vector of the number of characters in the name part of
each field, including the "=" separator and a trailing comma.
}
\description{
Calculate the number of characters in a field's name part
}
\keyword{internal}
