% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.nest_tree_impl}
\alias{.nest_tree_impl}
\title{Main implementation of tree nesting}
\usage{
.nest_tree_impl(
  x,
  id_col,
  parent_col,
  children_to,
  id_col_name,
  call = caller_env()
)
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{parent_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
identifies the parent id of each observation. Each value must either be
missing (for the root elements) or appear in the \code{id_col} column.}

\item{children_to}{(\code{character(1)}) The column name in which to store the
children.}

\item{id_col_name}{(\code{character(1)}) The name of the column that uniquely
identifies each observation.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A nested data frame with children column and parent column removed.
}
\description{
Main implementation of tree nesting
}
\keyword{internal}
