% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.nest_tree_lvl}
\alias{.nest_tree_lvl}
\title{Compute tree level for each element based on id and parent relationships}
\usage{
.nest_tree_lvl(ids, parent_ids, id_col_name, call = caller_env())
}
\arguments{
\item{id_col_name}{(\code{character(1)}) The name of the column that uniquely
identifies each observation.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A list with \code{lvls} (integer vector of levels) and \code{max_lvl} (maximum
level).
}
\description{
Compute tree level for each element based on id and parent relationships
}
\keyword{internal}
