% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.normalize_child_col_name}
\alias{.normalize_child_col_name}
\title{Resolve \code{child_col} to a column name string}
\usage{
.normalize_child_col_name(child_col, x, id_col, call = caller_env())
}
\arguments{
\item{child_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
contains the children of an observation. This column must be a list where
each element is either \code{NULL} or a data frame with the same columns as \code{x}.}

\item{x}{(\code{data.frame}) The data frame to unnest.}

\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{child_col} resolved to a \code{character(1)} column name.
}
\description{
Resolve \code{child_col} to a column name string
}
\keyword{internal}
