% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.normalize_id_col_name}
\alias{.normalize_id_col_name}
\title{Resolve \code{id_col} to a column name string}
\usage{
.normalize_id_col_name(id_col, x, call = caller_env())
}
\arguments{
\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{x}{(\code{data.frame}) The data frame to unnest.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{id_col} resolved to a \code{character(1)} column name.
}
\description{
Resolve \code{id_col} to a column name string
}
\keyword{internal}
