% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.normalize_parent_col}
\alias{.normalize_parent_col}
\title{Normalize and check the parent column}
\usage{
.normalize_parent_col(parent_col, x, id_col, id_col_name, call = caller_env())
}
\arguments{
\item{parent_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
identifies the parent id of each observation. Each value must either be
missing (for the root elements) or appear in the \code{id_col} column.}

\item{x}{(\code{any}) The object to check.}

\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{id_col_name}{(\code{character(1)}) The name of the column that uniquely
identifies each observation.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The integer index of the parent column.
}
\description{
Normalize and check the parent column
}
\keyword{internal}
