% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.pad}
\alias{.pad}
\title{Pad a character vector with a specified number of spaces}
\usage{
.pad(x, n)
}
\arguments{
\item{x}{(\code{character}) A vector to pad.}

\item{n}{(\code{integer}) The number of spaces to use for padding.}
}
\value{
A character vector with each element of \code{x} prepended with \code{n}
spaces. If an element of \code{x} contains newlines, each line will be prepended
with \code{n} spaces.
}
\description{
Pad a character vector with a specified number of spaces
}
\keyword{internal}
