% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_header_objects}
\alias{.parse_header_objects}
\title{Parse a named list of header objects from an OpenAPI spec}
\usage{
.parse_header_objects(header_objects, openapi_spec)
}
\arguments{
\item{header_objects}{(\code{list}) A named list of header objects from an
OpenAPI spec.}

\item{openapi_spec}{(\code{list}) The parsed OpenAPI specification.}
}
\value{
(\code{tbl_df}) A tibble of parsed header objects.
}
\description{
Parse a named list of header objects from an OpenAPI spec
}
\keyword{internal}
