% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_media_type_object}
\alias{.parse_media_type_object}
\title{Parse a single media type object from an OpenAPI spec}
\usage{
.parse_media_type_object(media_type_object, openapi_spec)
}
\arguments{
\item{media_type_object}{(\code{list}) A single media type object from an OpenAPI
spec.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A tibblify spec derived from the media type object's schema.
}
\description{
Parse a single media type object from an OpenAPI spec
}
\keyword{internal}
