% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_media_type_objects}
\alias{.parse_media_type_objects}
\title{Parse a named list of media type objects from an OpenAPI spec}
\usage{
.parse_media_type_objects(media_type_objects, openapi_spec)
}
\arguments{
\item{media_type_objects}{(\code{list}) A named list of media type objects from an
OpenAPI spec.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
(\code{tbl_df}) A tibble with \code{media_type} and \code{spec} columns.
}
\description{
Parse a named list of media type objects from an OpenAPI spec
}
\keyword{internal}
