% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_operation_object}
\alias{.parse_operation_object}
\title{Parse an operation object from an OpenAPI spec}
\usage{
.parse_operation_object(operation_object, openapi_spec)
}
\arguments{
\item{operation_object}{(\code{list}) An operation object from an OpenAPI spec, as
defined in the \href{https://spec.openapis.org/oas/v3.1.0#operation-object}{Operation Object}.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A one-row tibble describing the operation.
}
\description{
Parse an operation object from an OpenAPI spec
}
\keyword{internal}
