% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_path_item_object}
\alias{.parse_path_item_object}
\title{Parse a path item object from an OpenAPI spec}
\usage{
.parse_path_item_object(path_item_object, openapi_spec)
}
\arguments{
\item{path_item_object}{(\code{list}) A path item object from an OpenAPI spec,
as defined in the
\href{https://spec.openapis.org/oas/v3.1.0#path-item-object}{Path Item Object}.}

\item{openapi_spec}{(\code{list}) The parsed OpenAPI specification.}
}
\value{
A tibble of parsed operations with a \code{global_parameters} column.
}
\description{
Parse a path item object from an OpenAPI spec
}
\keyword{internal}
