% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_path_item_objects}
\alias{.parse_path_item_objects}
\title{Parse all path item objects from an OpenAPI spec}
\usage{
.parse_path_item_objects(openapi_spec)
}
\arguments{
\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A named list of tibbles, one per path, containing parsed operations.
}
\description{
Parse all path item objects from an OpenAPI spec
}
\keyword{internal}
