% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_request_body}
\alias{.parse_request_body}
\title{Parse a request body object from an OpenAPI spec}
\usage{
.parse_request_body(request_body, openapi_spec)
}
\arguments{
\item{request_body}{(\code{list} or \code{NULL}) A request body object from an OpenAPI
spec, as defined in the \href{https://spec.openapis.org/oas/v3.1.0#request-body-object}{Request Body Object}.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
(\code{tspec_row} or \code{NULL}) A parsed request body row spec, or \code{NULL}
if \code{request_body} is \code{NULL}.
}
\description{
Parse a request body object from an OpenAPI spec
}
\keyword{internal}
