% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.paste_multiline}
\alias{.paste_multiline}
\title{Pad and collapse a character vector across multiple lines}
\usage{
.paste_multiline(x)
}
\arguments{
\item{x}{(\code{character}) A vector to pad and collapse.}
}
\value{
A single string with the elements of \code{x} padded and collapsed across
multiple lines, with a leading newline and a trailing comma on each line.
}
\description{
Pad and collapse a character vector across multiple lines
}
\keyword{internal}
