% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.path_to_string}
\alias{.path_to_string}
\title{Convert a path object to a printable string}
\usage{
.path_to_string(path)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}
}
\value{
(\code{character(1)}) A string path such as \code{"x$a[[1]]"}.
}
\description{
Convert a path object to a printable string
}
\keyword{internal}
