% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.prep_complex_field}
\alias{.prep_complex_field}
\title{Prepare a complex field}
\usage{
.prep_complex_field(key, spec_fields)
}
\arguments{
\item{key}{(\code{character(1)}) The key of the field.}

\item{spec_fields}{(\code{list}) Fields from spec$fields (or a subset thereof).}
}
\value{
A prepared field specification.
}
\description{
Prepare a complex field
}
\keyword{internal}
