% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.prep_nested_keys}
\alias{.prep_nested_keys}
\title{Prepare nested keys in a tibblify specification}
\usage{
.prep_nested_keys(spec_fields, coll_locations)
}
\arguments{
\item{spec_fields}{(\code{list}) Fields from spec$fields (or a subset thereof).}

\item{coll_locations}{(\code{integer}) A numeric vector of locations.}
}
\value{
A list with prepared fields, keys, and locations.
}
\description{
Prepare nested keys in a tibblify specification
}
\keyword{internal}
