% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.prep_simple_field}
\alias{.prep_simple_field}
\title{Prepare a simple field}
\usage{
.prep_simple_field(x)
}
\arguments{
\item{x}{(\code{tib_collector}) A field specification.}
}
\value{
A prepared field specification.
}
\description{
Prepare a simple field
}
\keyword{internal}
