% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspec.R
\name{.prep_spec_fields}
\alias{.prep_spec_fields}
\title{Flatten, validate, and auto-name field specifications}
\usage{
.prep_spec_fields(.fields, .error_call)
}
\arguments{
\item{.fields}{(\code{list}) A list of field specifications, typically created by
\verb{tib_*()}.}

\item{.error_call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A named list of validated field specifications.
}
\description{
Flatten, validate, and auto-name field specifications
}
\keyword{internal}
