% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.prep_tib_fill}
\alias{.prep_tib_fill}
\title{Prepare the fill value for a vector field}
\usage{
.prep_tib_fill(fill, names_to, col_names)
}
\arguments{
\item{fill}{(\code{vector}) The fill value.}

\item{names_to}{(\code{character(1)}) Name of the names column.}

\item{col_names}{(\code{character}) Column names.}
}
\value{
The existing fill value, or a tibble representing the fill.
}
\description{
Prepare the fill value for a vector field
}
\keyword{internal}
