% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.prep_tib_list_of_ptype}
\alias{.prep_tib_list_of_ptype}
\title{Prepare the ptype list for a vector field}
\usage{
.prep_tib_list_of_ptype(ptype, names_to, col_names)
}
\arguments{
\item{ptype}{(\code{vector(0)}) The target ptype.}

\item{names_to}{(\code{character(1)}) Name of the names column.}

\item{col_names}{(\code{character}) Column names.}
}
\value{
A tibble representing the ptype.
}
\description{
Prepare the ptype list for a vector field
}
\keyword{internal}
