% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.prep_tib_scalar}
\alias{.prep_tib_scalar}
\title{Prepare a scalar field}
\usage{
.prep_tib_scalar(x)
}
\arguments{
\item{x}{(\code{tib_collector}) A field specification.}
}
\value{
A prepared scalar field specification.
}
\description{
Prepare a scalar field
}
\keyword{internal}
