% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.prep_tib_vector}
\alias{.prep_tib_vector}
\title{Prepare a vector field}
\usage{
.prep_tib_vector(x)
}
\arguments{
\item{x}{(\code{tib_collector}) A field specification.}
}
\value{
A prepared vector field specification.
}
\description{
Prepare a vector field
}
\keyword{internal}
