% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.prep_transform}
\alias{.prep_transform}
\title{Prepare transform function}
\usage{
.prep_transform(f, .call, arg = ".transform")
}
\arguments{
\item{f}{(\code{function} or \code{NULL}) The function to prepare.}

\item{.call}{(\code{environment}) The environment to use for error messages.}

\item{arg}{(\code{character(1)}) The argument name.}
}
\value{
(\code{function} or \code{NULL}) The prepared function.
}
\description{
Prepare transform function
}
\keyword{internal}
