% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.process_complex_fields}
\alias{.process_complex_fields}
\title{Process complex fields (depth > 1)}
\usage{
.process_complex_fields(spec_fields, first_keys)
}
\arguments{
\item{spec_fields}{(\code{list}) Fields from spec$fields (or a subset thereof).}

\item{first_keys}{(\code{character}) First keys of the complex fields.}
}
\value{
A list of prepared fields.
}
\description{
Process complex fields (depth > 1)
}
\keyword{internal}
