% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.process_simple_fields}
\alias{.process_simple_fields}
\title{Process simple fields (depth 1)}
\usage{
.process_simple_fields(spec_fields)
}
\arguments{
\item{spec_fields}{(\code{list}) Fields from spec$fields (or a subset thereof).}
}
\value{
A list of prepared fields.
}
\description{
Process simple fields (depth 1)
}
\keyword{internal}
