% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.read_spec}
\alias{.read_spec}
\title{Read an OpenAPI spec from a file, connection, or list}
\usage{
.read_spec(file, arg = caller_arg(file), call = caller_env())
}
\arguments{
\item{file}{(\code{character(1)}) A path to a file, a connection, or literal data.}

\item{arg}{(\code{character(1)}) An argument name. This name will be mentioned in
error messages as the input that is at the origin of a problem.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{list}) The parsed spec.
}
\description{
Read an OpenAPI spec from a file, connection, or list
}
\keyword{internal}
