% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi-read_spec.R
\name{.read_spec_impl}
\alias{.read_spec_impl}
\alias{.read_spec_impl.list}
\alias{.read_spec_impl.connection}
\alias{.read_spec_impl.character}
\alias{.read_spec_impl.default}
\title{Read an OpenAPI spec or schema from a file, connection, or list}
\usage{
.read_spec_impl(file, arg = caller_arg(file), call = caller_env())

\method{.read_spec_impl}{list}(file, ...)

\method{.read_spec_impl}{connection}(file, ...)

\method{.read_spec_impl}{character}(file, arg = caller_arg(file), call = caller_env())

\method{.read_spec_impl}{default}(file, arg = caller_arg(file), call = caller_env())
}
\arguments{
\item{file}{(\code{character(1)}) A path to a file, a connection, or literal data.}

\item{arg}{(\code{character(1)}) An argument name. This name will be mentioned in
error messages as the input that is at the origin of a problem.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{list}) The parsed spec or schema.
}
\description{
Read an OpenAPI spec or schema from a file, connection, or list
}
\keyword{internal}
