% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.reduce_ptype}
\alias{.reduce_ptype}
\title{Reduce per-level data frames to their combined ptype}
\usage{
.reduce_ptype(snapshots, call)
}
\arguments{
\item{snapshots}{(\code{list}) The object returned by \code{.walk_tree_levels()}.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A 0-row data frame representing the combined type of all levels.
}
\description{
Reduce per-level data frames to their combined ptype
}
\keyword{internal}
