% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.remove_first_key}
\alias{.remove_first_key}
\title{Remove the first key from a field}
\usage{
.remove_first_key(x)
}
\arguments{
\item{x}{(\code{tib_collector}) A field specification.}
}
\value{
The field specification with the first key removed.
}
\description{
Remove the first key from a field
}
\keyword{internal}
