% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.schema_has_ref}
\alias{.schema_has_ref}
\title{Check whether any nested list element has a \verb{$ref} key}
\usage{
.schema_has_ref(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
(\code{logical(1)}) \code{TRUE} if any element has a \verb{$ref} key.
}
\description{
Check whether any nested list element has a \verb{$ref} key
}
\keyword{internal}
