% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_tspec_object.R
\name{.set_spec}
\alias{.set_spec}
\title{Set the tibblify specification attribute}
\usage{
.set_spec(x, spec)
}
\arguments{
\item{x}{(\code{list}) A nested list.}

\item{spec}{(\code{tspec}) A specification of how to convert \code{x}. Generated with
\code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}}, \code{\link[=tspec_object]{tspec_object()}}, \code{\link[=tspec_recursive]{tspec_recursive()}}, or (if
\code{NULL}, the default), \code{\link[=guess_tspec]{guess_tspec()}}.}
}
\value{
The object \code{x} with the \code{tib_spec} attribute set.
}
\description{
Set the tibblify specification attribute
}
\keyword{internal}
