% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared_params.R
\name{.shared-params-openapi}
\alias{.shared-params-openapi}
\title{Shared OpenAPI parsing parameters}
\arguments{
\item{header_objects}{(\code{list}) A named list of header objects from an
OpenAPI spec.}

\item{media_type_object}{(\code{list}) A single media type object from an
OpenAPI spec.}

\item{media_type_objects}{(\code{list}) A named list of media type objects from
an OpenAPI spec.}

\item{openapi_spec}{(\code{list}) The parsed OpenAPI specification.}

\item{operation_object}{(\code{list}) An operation object from an OpenAPI spec,
as defined in the
\href{https://spec.openapis.org/oas/v3.1.0#operation-object}{Operation Object}.}

\item{parameters}{(\code{list} or \code{NULL}) A list of parameter objects from an
OpenAPI spec.}

\item{path_item_object}{(\code{list}) A path item object from an OpenAPI spec,
as defined in the
\href{https://spec.openapis.org/oas/v3.1.0#path-item-object}{Path Item Object}.}

\item{request_body}{(\code{list} or \code{NULL}) A request body object from an
OpenAPI spec.}

\item{response_object}{(\code{list}) A response object from an OpenAPI spec.}

\item{responses_object}{(\code{list}) A responses object from an OpenAPI spec,
mapping status codes to response objects.}

\item{schema}{(\code{list}) A JSON schema object, typically from
\code{openapi_spec$components$schemas} or inline within the spec.}
}
\description{
These parameters are used in multiple internal functions in
\code{R/parse_openapi.R}. They are defined here to make them easier to import
and to find.
}
\keyword{internal}
