% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.shared-params-spec_prep}
\alias{.shared-params-spec_prep}
\title{Shared spec_prep parameters}
\arguments{
\item{coll_locations}{(\code{integer}) A numeric vector of locations.}

\item{field_spec}{(\code{tib_collector}) A field specification.}

\item{fill}{(\code{vector}) The fill value.}

\item{first_keys}{(\code{character}) First keys of the complex fields.}

\item{key}{(\code{character(1)}) The key of the field.}

\item{spec_fields}{(\code{list}) Fields from spec$fields (or a subset thereof).}

\item{x}{(\code{tib_collector}) A field specification.}

\item{ptype}{(\code{vector(0)}) The target ptype.}

\item{names_to}{(\code{character(1)}) Name of the names column.}

\item{col_names}{(\code{character}) Column names.}
}
\description{
These parameters are used in multiple \code{\link[=.spec_prep]{.spec_prep()}}-related functions. They
are defined here to make them easier to import and to find. This break-out is
for parameters that may differ from other functions that use the same
parameter names.
}
\keyword{internal}
