% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared_params.R
\name{.shared-params-tib}
\alias{.shared-params-tib}
\title{Shared tib_spec parameters}
\arguments{
\item{.input_form}{A string that describes the structure of the field. Can be
one of:
\itemize{
\item \code{"vector"}: The field is a vector, e.g. \code{c(1, 2, 3)}.
\item \code{"scalar_list"}: The field is a list of scalars, e.g. \code{list(1, 2, 3)}.
\item \code{"object"}: The field is a named list of scalars, e.g. \code{list(a = 1, b = 2, c = 3)}.
}}

\item{.names_to}{What to do with the inner names of the object. Can be one
of:
\itemize{
\item \code{NULL}: the default. The inner names of the field are not used.
\item A string: This can only be used if 1) the input form is \code{"object"}
or \code{"vector"} and 2) \code{.values_to} is a string. The inner names of the field
will populate the specified column in the field's tibble.
}}
}
\description{
These parameters are used in multiple \verb{tib_*()} functions. They are defined
here to make them easier to import and to find, and to make sure the
deprecated forms get documented identically when these definitions are
imported. This break-out is for parameters that differ between \verb{tib_*()}
functions and other functions that use the same parameters.
}
\keyword{internal}
