% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared_params.R
\name{.shared-params-tree}
\alias{.shared-params-tree}
\title{Shared (un)nest_tree parameters}
\arguments{
\item{col}{(\code{character}, \code{integer}, or \code{symbol}) Defused R code describing a
selection according to the tidyselect syntax.}

\item{col_arg}{(\code{character(1)}) The name of the \code{col} argument, used for
error messages.}

\item{ids}{(\code{character} or \code{integer}) The potential child ids to compare
against.}

\item{parent_ids}{(\code{character} or \code{integer}) The parent ids to check.}
}
\description{
These parameters are used in multiple \code{nest_tree()} or \code{unnest_tree()}
helpers. They are defined here to make them easier to import and to find.
This break-out is for parameters that differ between \verb{(un)nest_tree()}
helpers and other functions that might use the same parameters.
}
\keyword{internal}
