% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.should_force_names}
\alias{.should_force_names}
\title{Determine whether to print names for tibblifying}
\usage{
.should_force_names()
}
\value{
A logical value indicating whether to print names for tibblifying,
based on the \code{tibblify.print_names} option.
}
\description{
Determine whether to print names for tibblifying
}
\keyword{internal}
