% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.sort_spec_by_first_key}
\alias{.sort_spec_by_first_key}
\title{Sort specification by the first key}
\usage{
.sort_spec_by_first_key(spec_fields, coll_locations)
}
\arguments{
\item{spec_fields}{(\code{list}) Fields from spec$fields (or a subset thereof).}

\item{coll_locations}{(\code{integer}) A numeric vector of locations.}
}
\value{
A list with sorted components.
}
\description{
Sort specification by the first key
}
\keyword{internal}
