% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspec.R
\name{.spec_auto_name_fields}
\alias{.spec_auto_name_fields}
\title{Auto-name fields based on their key attribute}
\usage{
.spec_auto_name_fields(.fields, .error_call)
}
\arguments{
\item{.fields}{(\code{list}) A list of field specifications, typically created by
\verb{tib_*()}.}

\item{.error_call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A named list of field specifications.
}
\description{
Auto-name fields based on their key attribute
}
\keyword{internal}
