% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_inform_unspecified.R
\name{.spec_inform_unspecified}
\alias{.spec_inform_unspecified}
\title{Inform about or error for unspecified fields}
\usage{
.spec_inform_unspecified(spec, unspecified = "inform", call = caller_env())
}
\arguments{
\item{spec}{(\code{tspec}) A specification of how to convert \code{x}. Generated with
\code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}}, \code{\link[=tspec_object]{tspec_object()}}, \code{\link[=tspec_recursive]{tspec_recursive()}}, or (if
\code{NULL}, the default), \code{\link[=guess_tspec]{guess_tspec()}}.}

\item{unspecified}{(\code{character(1)}) What to do with unspecified fields. Can
be one of
\itemize{
\item \code{"error"}: Throw an error.
\item \code{"inform"}: Inform the user.
\item \code{"drop"}: Do not parse these fields.
\item \code{"list"}: Parse an unspecified field into a list as with \code{\link[=tib_variant]{tib_variant()}}.
}}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The original spec, invisibly.
}
\description{
Inform about or error for unspecified fields
}
\keyword{internal}
