% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep.R
\name{.spec_prep_nonrecursive}
\alias{.spec_prep_nonrecursive}
\title{Prepare a non-recursive tibblify specification}
\usage{
.spec_prep_nonrecursive(spec)
}
\arguments{
\item{spec}{(\code{tspec}) A specification of how to convert \code{x}. Generated with
\code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}}, \code{\link[=tspec_object]{tspec_object()}}, \code{\link[=tspec_recursive]{tspec_recursive()}}, or
\code{\link[=guess_tspec]{guess_tspec()}}. If \code{spec} is \code{NULL} (the default), \code{guess_tspec(x, inform_unspecified = TRUE)} will be used to guess the \code{spec}.}
}
\value{
A prepared tibblify specification.
}
\description{
Prepare a non-recursive tibblify specification
}
\keyword{internal}
