% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep_unspecified.R
\name{.spec_replace_unspecified_impl}
\alias{.spec_replace_unspecified_impl}
\title{Replace unspecified fields in the specification}
\usage{
.spec_replace_unspecified_impl(spec_fields, unspecified)
}
\arguments{
\item{spec_fields}{(\code{list}) Fields from spec$fields (or a subset thereof).}

\item{unspecified}{(\code{character(1)}) What to do with unspecified fields. Can
be one of
\itemize{
\item \code{"error"}: Throw an error.
\item \code{"inform"}: Inform the user.
\item \code{"drop"}: Do not parse these fields.
\item \code{"list"}: Parse an unspecified field into a list as with \code{\link[=tib_variant]{tib_variant()}}.
}}
}
\value{
A modified list of \code{tib_collector} objects.
}
\description{
Replace unspecified fields in the specification
}
\keyword{internal}
