% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep_unspecified.R
\name{.spec_replace_unspecified_type}
\alias{.spec_replace_unspecified_type}
\title{Replace tib_unspecified fields in the specification}
\usage{
.spec_replace_unspecified_type(field_spec, unspecified)
}
\arguments{
\item{field_spec}{(\code{tib_collector}) A field specification.}

\item{unspecified}{(\code{character(1)}) What to do with \code{\link[=tib_unspecified]{tib_unspecified()}}
fields. Can be one of
\itemize{
\item \code{"error"}: Throw an error.
\item \code{"inform"}: Inform the user then parse as with \code{\link[=tib_variant]{tib_variant()}}.
\item \code{"drop"}: Do not parse these fields.
\item \code{"list"}: Parse unspecified fields into lists as with \code{\link[=tib_variant]{tib_variant()}}.
}}
}
\value{
A modified \code{tib_collector}.
}
\description{
Replace tib_unspecified fields in the specification
}
\keyword{internal}
