% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.stabilize_names_to}
\alias{.stabilize_names_to}
\title{Stabilize names_to argument}
\usage{
.stabilize_names_to(.names_to, .values_to, .input_form, .call)
}
\arguments{
\item{.names_to}{(\code{character(1)} or \code{NULL}) What to do with the inner names
of the object. Can be one of:
\itemize{
\item \code{NULL}: the default. The inner names of the field are not used.
\item A string: Use only if the input form is \code{"object"} or \code{"vector"}, and \code{.values_to} is a string. The inner names of the field will populate the
specified column in the field's tibble.
}}

\item{.values_to}{(\code{character(1)} or \code{NULL}) For \code{NULL} (the default), the
field is converted to a \code{.ptype} vector. If a string is provided, the field
is converted to a tibble and the values go into the specified column.}

\item{.input_form}{(\code{character(1)}) The structure of the input field. Can be
one of:
\itemize{
\item \code{"vector"}: The field is a vector, e.g. \code{c(1, 2, 3)}.
\item \code{"scalar_list"}: The field is a list of scalars, e.g. \code{list(1, 2, 3)}.
\item \code{"object"}: The field is a named list of scalars, e.g.
\code{list(a = 1, b = 2, c = 3)}.
}}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{character(1)} or \code{NULL}) Validated \code{.names_to}.
}
\description{
Stabilize names_to argument
}
\keyword{internal}
