% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_tspec.R
\name{.stabilize_unpack_cols}
\alias{.stabilize_unpack_cols}
\title{Check which fields to unpack}
\usage{
.stabilize_unpack_cols(fields, spec, .call = caller_env())
}
\arguments{
\item{fields}{(\code{character} or \code{NULL}) The fields to unpack.}

\item{spec}{(\code{tspec}) A tibblify specification.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{character}) The names of the fields to unpack.
}
\description{
Check which fields to unpack
}
\keyword{internal}
