% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.stabilize_values_to}
\alias{.stabilize_values_to}
\title{Stabilize values_to argument}
\usage{
.stabilize_values_to(.values_to, .call)
}
\arguments{
\item{.values_to}{(\code{character(1)} or \code{NULL}) For \code{NULL} (the default), the
field is converted to a \code{.ptype} vector. If a string is provided, the field
is converted to a tibble and the values go into the specified column.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{character(1)} or \code{NULL}) Validated \code{.values_to}.
}
\description{
Stabilize values_to argument
}
\keyword{internal}
