% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_colmajor_null}
\alias{.stop_colmajor_null}
\title{Error for NULL in colmajor fields}
\usage{
.stop_colmajor_null(path)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}
}
\value{
Never returns; called for its side effect of throwing an error.
}
\description{
Error for NULL in colmajor fields
}
\keyword{internal}
