% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_colmajor_wrong_size_element}
\alias{.stop_colmajor_wrong_size_element}
\title{Error for inconsistent colmajor field sizes}
\usage{
.stop_colmajor_wrong_size_element(path, size_act, path_exp, size_exp)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}

\item{size_act}{(\code{integer(1)}) The observed size of a field.}

\item{path_exp}{(\code{list}) The path of the field used as the reference in size
mismatch errors.}

\item{size_exp}{(\code{integer(1)}) The expected size of a field.}
}
\value{
\code{NULL} (invisibly). Called for its side effect of throwing an error.
}
\description{
Error for inconsistent colmajor field sizes
}
\keyword{internal}
