% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_duplicate_name}
\alias{.stop_duplicate_name}
\title{Error for duplicate names}
\usage{
.stop_duplicate_name(path, name)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}

\item{name}{(\code{character(1)}) The name of the field.}
}
\value{
Never returns; called for its side effect of throwing an error.
}
\description{
Error for duplicate names
}
\keyword{internal}
