% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_empty_name}
\alias{.stop_empty_name}
\title{Error for empty names}
\usage{
.stop_empty_name(path, index)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}

\item{index}{(\code{integer(1)}) A zero-based location in a path.}
}
\value{
Never returns; called for its side effect of throwing an error.
}
\description{
Error for empty names
}
\keyword{internal}
