% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_names_is_null}
\alias{.stop_names_is_null}
\title{Error for unnamed object}
\usage{
.stop_names_is_null(path)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}
}
\value{
Never returns; called for its side effect of throwing an error.
}
\description{
Error for unnamed object
}
\keyword{internal}
