% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_non_list_element}
\alias{.stop_non_list_element}
\title{Error for non-list element}
\usage{
.stop_non_list_element(path, x)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}

\item{x}{(\code{any}) The object to check.}
}
\value{
Never returns; called for its side effect of throwing an error.
}
\description{
Error for non-list element
}
\keyword{internal}
