% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_object_vector_names_is_null}
\alias{.stop_object_vector_names_is_null}
\title{Error for unnamed object vector}
\usage{
.stop_object_vector_names_is_null(path)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}
}
\value{
\code{NULL} (invisibly). Called for its side effect of throwing an error.
}
\description{
Error for unnamed object vector
}
\keyword{internal}
