% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_scalar}
\alias{.stop_scalar}
\title{Error for non-scalar field}
\usage{
.stop_scalar(path, size_act)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}

\item{size_act}{(\code{integer(1)}) The observed size of a field.}
}
\value{
\code{NULL} (invisibly). Called for its side effect of throwing an error.
}
\description{
Error for non-scalar field
}
\keyword{internal}
