% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_vector_wrong_size_element}
\alias{.stop_vector_wrong_size_element}
\title{Error for wrong-sized vector element}
\usage{
.stop_vector_wrong_size_element(path, input_form, x)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}

\item{input_form}{(\code{character(1)}) The input form string used in error
messages.}

\item{x}{(\code{any}) The object to check.}
}
\value{
\code{NULL} (invisibly). Called for its side effect of throwing an error.
}
\description{
Error for wrong-sized vector element
}
\keyword{internal}
