% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.tib_collector}
\alias{.tib_collector}
\title{Create a tib collector}
\usage{
.tib_collector(
  .key,
  .type,
  ...,
  .required = TRUE,
  .class = NULL,
  .transform = NULL,
  .elt_transform = NULL,
  .call = caller_env()
)
}
\arguments{
\item{.key}{(\code{character}) The path of names to the field in the object.}

\item{.type}{(\code{character(1)}) The type of the collector.}

\item{.required}{(\verb{logical(1))} Throw an error if the field does not exist?}

\item{.class}{(\code{character} or \code{NULL}) Additional classes for the collector.}

\item{.transform}{(\code{function} or \code{NULL}) A function to apply to the whole
vector after casting to \code{.ptype_inner}.}

\item{.elt_transform}{(\code{function} or \code{NULL}) A function to apply to each
element before casting to \code{.ptype_inner}.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{tib_collector}) A tibblify collector.
}
\description{
Create a tib collector
}
\keyword{internal}
