% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine}
\alias{.tib_combine}
\title{Combine a list of tib fields}
\usage{
.tib_combine(tib_list, name, .call)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{tib_collector}) A combined tibblify field collector.
}
\description{
Combine a list of tib fields
}
\keyword{internal}
