% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine_df}
\alias{.tib_combine_df}
\title{Combine dataframe tib fields}
\usage{
.tib_combine_df(tib_list, type, key, required, .call)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}

\item{type}{(\code{character(1)}) The target tib type ("row" or "df").}

\item{key}{(\code{character(1)}) The key of the field.}

\item{required}{(\code{logical(1)}) Whether the field is required.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{tib_row} or \code{tib_df}) A tibblify dataframe field collector.
}
\description{
Combine dataframe tib fields
}
\keyword{internal}
