% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine_fill}
\alias{.tib_combine_fill}
\title{Combine fill values from a list of tibs}
\usage{
.tib_combine_fill(tib_list, type, ptype, .call)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}

\item{type}{(\code{character(1)}) The target tib type (scalar, vector, etc.).}

\item{ptype}{(\code{vector(0)}) The target ptype.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{vector}) The combined fill value.
}
\description{
Combine fill values from a list of tibs
}
\keyword{internal}
