% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine_input_form}
\alias{.tib_combine_input_form}
\title{Combine input forms from a list of tibs}
\usage{
.tib_combine_input_form(tib_list, .call)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{character(1)}) The combined input form.
}
\description{
Combine input forms from a list of tibs
}
\keyword{internal}
