% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine_key}
\alias{.tib_combine_key}
\title{Combine keys from a list of tibs}
\usage{
.tib_combine_key(tib_list, name, .call)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{character(1)}) The combined key.
}
\description{
Combine keys from a list of tibs
}
\keyword{internal}
