% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine_names_col}
\alias{.tib_combine_names_col}
\title{Combine names_col value from a list of tibs}
\usage{
.tib_combine_names_col(tib_list, .call)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{character(1)} or \code{NULL}) The combined names column.
}
\description{
Combine names_col value from a list of tibs
}
\keyword{internal}
