% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine_ptype}
\alias{.tib_combine_ptype}
\title{Combine ptypes from a list of tibs}
\usage{
.tib_combine_ptype(tib_list, .call)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{vector(0)}) The combined ptype.
}
\description{
Combine ptypes from a list of tibs
}
\keyword{internal}
