% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine_required}
\alias{.tib_combine_required}
\title{Combine required status from a list of tibs}
\usage{
.tib_combine_required(tib_list)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}
}
\value{
(\code{logical(1)}) The combined required status.
}
\description{
Combine required status from a list of tibs
}
\keyword{internal}
