% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine_specified}
\alias{.tib_combine_specified}
\title{Combine various specified tib field types}
\usage{
.tib_combine_specified(tib_list, type, key, required, .call)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}

\item{type}{(\code{character(1)}) The target tib type ("variant", "scalar", or
"vector").}

\item{key}{(\code{character(1)}) The key of the field.}

\item{required}{(\code{logical(1)}) Whether the field is required.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{tib_variant}, \code{tib_scalar}, or \code{tib_vector}) A tibblify field
collector.
}
\description{
Combine various specified tib field types
}
\keyword{internal}
