% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine_transform}
\alias{.tib_combine_transform}
\title{Combine transform functions from a list of tibs}
\usage{
.tib_combine_transform(tib_list, .call)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{function} or \code{NULL}) The combined transform function.
}
\description{
Combine transform functions from a list of tibs
}
\keyword{internal}
