% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tib_combine_vector}
\alias{.tib_combine_vector}
\title{Combine vector tib fields}
\usage{
.tib_combine_vector(tib_list, key, ptype, required, fill, transform, .call)
}
\arguments{
\item{tib_list}{(\code{list}) A list of tib fields.}

\item{key}{(\code{character(1)}) The key of the field.}

\item{ptype}{(\code{vector(0)}) The target ptype.}

\item{required}{(\code{logical(1)}) Whether the field is required.}

\item{fill}{(\code{vector}) The fill value.}

\item{transform}{(\code{function} or \code{NULL}) The transform function.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{tib_vector}) A tibblify vector field collector.
}
\description{
Combine vector tib fields
}
\keyword{internal}
