% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.tib_from_schema_object}
\alias{.tib_from_schema_object}
\title{Parse an object-type OpenAPI schema and create a tib field}
\usage{
.tib_from_schema_object(schema, name, openapi_spec)
}
\arguments{
\item{schema}{(\code{list}) A JSON schema object, as defined in the \href{https://spec.openapis.org/oas/v3.1.0#schema-object}{Schema Object}, typically from
\code{openapi_spec$components$schemas} or inline within the spec.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A \code{\link[=tib_row]{tib_row()}} field spec with fields extracted from schema
properties.
}
\description{
Parse an object-type OpenAPI schema and create a tib field
}
\keyword{internal}
