% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_utils.R
\name{.tib_ptype}
\alias{.tib_ptype}
\title{Get the ptype of an object}
\usage{
.tib_ptype(x)
}
\arguments{
\item{x}{(\code{any}) The object to check.}
}
\value{
The ptype of \code{x}, with \code{POSIXlt} coerced to \code{POSIXct}.
}
\description{
Get the ptype of an object
}
\keyword{internal}
