% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object_utils.R
\name{.tib_scalar_or_vector_spec}
\alias{.tib_scalar_or_vector_spec}
\title{Create a scalar or vector tib spec}
\usage{
.tib_scalar_or_vector_spec(name, ptype, is_scalar)
}
\arguments{
\item{name}{(\code{character(1)}) The name of the field.}

\item{ptype}{(\code{vector(0)}) A prototype of the desired output type of the
field.}

\item{is_scalar}{(\code{logical(1)}) If \code{TRUE}, return a \code{\link[=tib_scalar]{tib_scalar()}} spec,
otherwise a \code{\link[=tib_vector]{tib_vector()}} spec.}
}
\value{
A \code{\link[=tib_scalar]{tib_scalar()}} or \code{\link[=tib_vector]{tib_vector()}} spec.
}
\description{
Create a scalar or vector tib spec
}
\keyword{internal}
