% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_utils.R
\name{.tib_type_of}
\alias{.tib_type_of}
\title{Determine the tib type of an object}
\usage{
.tib_type_of(x, name, other)
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{other}{(\code{logical(1)}) If \code{TRUE}, return \code{"other"} for unrecognized
types rather than throwing an error.}
}
\value{
One of \code{"df"}, \code{"list"}, \code{"vector"}, or \code{"other"}.
}
\description{
Determine the tib type of an object
}
\keyword{internal}
