% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object_utils.R
\name{.tib_vector_input_form}
\alias{.tib_vector_input_form}
\title{Determine the vector input form of a value}
\usage{
.tib_vector_input_form(value)
}
\arguments{
\item{value}{(\code{list}) An object list whose fields will be guessed.}
}
\value{
\code{"object"} if \code{value} is named, \code{"scalar_list"} otherwise.
}
\description{
Determine the vector input form of a value
}
\keyword{internal}
