% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tibblify_abort}
\alias{.tibblify_abort}
\title{Throw a tibblify internal error}
\usage{
.tibblify_abort(..., .envir = caller_env())
}
\arguments{
\item{...}{Arguments passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}.}

\item{.envir}{(\code{environment}) The environment used to evaluate cli fields.}
}
\value{
\code{NULL} (invisibly). Called for its side effect of throwing an error.
}
\description{
Throw a tibblify internal error
}
\keyword{internal}
