% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibblify.R
\name{.tibblify_impl}
\alias{.tibblify_impl}
\title{Tibblify implementation}
\usage{
.tibblify_impl(x, spec, path)
}
\arguments{
\item{x}{(\code{list}) A nested list.}

\item{spec}{(\code{tspec}) A specification of how to convert \code{x}. Generated with
\code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}}, \code{\link[=tspec_object]{tspec_object()}}, \code{\link[=tspec_recursive]{tspec_recursive()}}, or
\code{\link[=guess_tspec]{guess_tspec()}}. If \code{spec} is \code{NULL} (the default), \code{guess_tspec(x, inform_unspecified = TRUE)} will be used to guess the \code{spec}.}

\item{path}{(\code{list}) The current path in the data structure.}
}
\value{
Either a tibble or a list, depending on the specification.
}
\description{
Tibblify implementation
}
\keyword{internal}
