% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.tibblify_width}
\alias{.tibblify_width}
\title{Get the width for tibblifying}
\usage{
.tibblify_width(width = NULL)
}
\arguments{
\item{width}{(\code{integer(1)}) The width (in number of characers) of text output
to generate.}
}
\value{
The width to use for tibblifying, either the provided \code{width} or the
value of the \code{width} option if \code{width} is \code{NULL}.
}
\description{
Get the width for tibblifying
}
\keyword{internal}
