% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibblify.R
\name{.try_tibblify_impl}
\alias{.try_tibblify_impl}
\title{Tibblify implementation with error handling}
\usage{
.try_tibblify_impl(x, spec, call = caller_env())
}
\arguments{
\item{x}{(\code{list}) A nested list.}

\item{spec}{(\code{tspec}) A specification of how to convert \code{x}. Generated with
\code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}}, \code{\link[=tspec_object]{tspec_object()}}, \code{\link[=tspec_recursive]{tspec_recursive()}}, or
\code{\link[=guess_tspec]{guess_tspec()}}. If \code{spec} is \code{NULL} (the default), \code{guess_tspec(x, inform_unspecified = TRUE)} will be used to guess the \code{spec}.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
Either a tibble or a list, depending on the specification.
}
\description{
Tibblify implementation with error handling
}
\keyword{internal}
