% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspec.R
\name{.tspec}
\alias{.tspec}
\title{Create a tibblify specification object}
\usage{
.tspec(
  .fields,
  .type,
  ...,
  .vector_allows_empty_list = FALSE,
  .error_call = caller_env()
)
}
\arguments{
\item{.fields}{(\code{list}) A list of field specifications, typically created by
\verb{tib_*()}.}

\item{.type}{(\code{character(1)}) The type of specification being created
(\code{"df"}, \code{"object"}, \code{"row"}, or \code{"recursive"}).}

\item{...}{Additional specification attributes passed to \code{\link[rlang:list2]{rlang::list2()}}.}

\item{.error_call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A \code{tspec} object with class \verb{tspec_<type>} and \code{tspec}.
}
\description{
Create a tibblify specification object
}
\keyword{internal}
