% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{.tspec_combine_field_list}
\alias{.tspec_combine_field_list}
\title{Combine fields from a list of specs}
\usage{
.tspec_combine_field_list(spec_list, .call)
}
\arguments{
\item{spec_list}{(\code{list}) A list of specifications.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{list}) A list of combined fields generated by \code{\link[=.tib_combine]{.tib_combine()}}.
}
\description{
Combine fields from a list of specs
}
\keyword{internal}
