% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.tspec_from_one_of}
\alias{.tspec_from_one_of}
\title{Combine \code{oneOf} sub-schemas into a tspec}
\usage{
.tspec_from_one_of(schema, openapi_spec)
}
\arguments{
\item{schema}{(\code{list}) A JSON schema object, as defined in the \href{https://spec.openapis.org/oas/v3.1.0#schema-object}{Schema Object}, typically from
\code{openapi_spec$components$schemas} or inline within the spec.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A \code{\link[=tspec_combine]{tspec_combine()}}, or \code{tib_variant("dummy")} if schemas are
incompatible, or \code{NULL} if \code{schema$oneOf} is empty.
}
\description{
Combine \code{oneOf} sub-schemas into a tspec
}
\keyword{internal}
