% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.tspec_from_schema_array}
\alias{.tspec_from_schema_array}
\title{Convert an array-type schema to a tibblify data frame spec}
\usage{
.tspec_from_schema_array(schema, openapi_spec)
}
\arguments{
\item{schema}{(\code{list}) A JSON schema object, as defined in the \href{https://spec.openapis.org/oas/v3.1.0#schema-object}{Schema Object}, typically from
\code{openapi_spec$components$schemas} or inline within the spec.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A \code{\link[=tspec_df]{tspec_df()}} spec with fields extracted from the array items
schema properties.
}
\description{
Convert an array-type schema to a tibblify data frame spec
}
\keyword{internal}
