% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.tspec_from_schema_object}
\alias{.tspec_from_schema_object}
\title{Convert an object-type schema to a tibblify row spec}
\usage{
.tspec_from_schema_object(schema, openapi_spec)
}
\arguments{
\item{schema}{(\code{list}) A JSON schema object, as defined in the \href{https://spec.openapis.org/oas/v3.1.0#schema-object}{Schema Object}, typically from
\code{openapi_spec$components$schemas} or inline within the spec.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A \code{\link[=tspec_row]{tspec_row()}} spec with fields extracted from schema properties.
}
\description{
Convert an object-type schema to a tibblify row spec
}
\keyword{internal}
