% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.tspec_from_schema_type}
\alias{.tspec_from_schema_type}
\title{Choose and apply a tspec_from_schema_type function}
\usage{
.tspec_from_schema_type(schema, openapi_spec)
}
\arguments{
\item{schema}{(\code{list}) A JSON schema object, as defined in the \href{https://spec.openapis.org/oas/v3.1.0#schema-object}{Schema Object}, typically from
\code{openapi_spec$components$schemas} or inline within the spec.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A tibblify spec (\code{\link[=tspec_row]{tspec_row()}}, \code{\link[=tspec_df]{tspec_df()}}, or a tib field).
}
\description{
Choose and apply a tspec_from_schema_type function
}
\keyword{internal}
