% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.unchop_children}
\alias{.unchop_children}
\title{Unclass and unchop a list of child data frames}
\usage{
.unchop_children(children, child_col, call = caller_env())
}
\arguments{
\item{children}{(\code{list}) List of child data frames or \code{NULL} elements.}

\item{child_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
contains the children of an observation. This column must be a list where
each element is either \code{NULL} or a data frame with the same columns as \code{x}.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{data.frame}) All non-\code{NULL} children combined into one data frame.
}
\description{
Unclass and unchop a list of child data frames
}
\keyword{internal}
