% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_tspec.R
\name{.unchop_fields}
\alias{.unchop_fields}
\title{Unchop and repair fields}
\usage{
.unchop_fields(fields, names_repair, names_clean, .call = caller_env())
}
\arguments{
\item{fields}{(\code{character} or \code{NULL}) The fields to unpack. If \code{fields} is
\code{NULL} (default), all fields are unpacked.}

\item{names_repair}{(\code{character(1)} or \code{function}) Passed to the \code{repair}
argument of \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} to check that the output data frame has
valid names. Must be one of the following options:
\itemize{
\item \code{"unique"} (the default) or \code{"unique_quiet"}: make sure names are unique
and not empty,
\item \code{"universal"} or \code{"universal_quiet"}: make the names unique and syntactic
\item \code{"check_unique"}: no name repair, but check they are unique,
\item a function: apply custom name repair.
}}

\item{names_clean}{(\code{function}) A one-argument function to clean names after
repairing. For example use \code{\link[=camel_case_to_snake_case]{camel_case_to_snake_case()}}.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{list}) A list of unchopped and repaired fields.
}
\description{
Unchop and repair fields
}
\keyword{internal}
