% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{.unclass_list_of}
\alias{.unclass_list_of}
\title{Unclass a list_of child data frame}
\usage{
.unclass_list_of(x, child_col, call = caller_env())
}
\arguments{
\item{x}{(\code{data.frame} or \code{NULL}) A single child data frame.}

\item{child_col}{(\code{character(1)}) Name of the column that may itself be a
\code{vctrs_list_of}.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{data.frame} or \code{NULL}) \code{x} with \code{vctrs_list_of} classes removed,
or \code{NULL} if \code{x} is \code{NULL}.
}
\description{
Unclass a list_of child data frame
}
\keyword{internal}
