% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_tspec.R
\name{.unpack_field_row}
\alias{.unpack_field_row}
\title{Unpack a row field}
\usage{
.unpack_field_row(field_spec, name, names_sep)
}
\arguments{
\item{field_spec}{(\code{tib_collector}) A tibblify field collector.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{names_sep}{(\code{character(1)} or \code{NULL}) If \code{NULL}, the default, the inner
names of fields are used. If a string, the outer and inner names are pasted
together, separated by \code{names_sep}.}
}
\value{
(\code{list}) A list of unpacked fields from the row spec.
}
\description{
Unpack a row field
}
\keyword{internal}
