% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_tspec.R
\name{.unpack_key}
\alias{.unpack_key}
\title{Update the key of an unpacked field}
\usage{
.unpack_key(field, key)
}
\arguments{
\item{field}{(\code{tib_collector}) The field spec being unpacked.}

\item{key}{(\code{character}) The key to prepend.}
}
\value{
(\code{tib_collector}) The field with updated key.
}
\description{
Update the key of an unpacked field
}
\keyword{internal}
