% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untibblify.R
\name{.untibblify_df}
\alias{.untibblify_df}
\title{Untibblify a data frame into a list of row lists}
\usage{
.untibblify_df(x, spec, call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{spec}{(\code{tspec} or \code{NULL}) A spec object describing the structure of
\code{x}.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A list with one named list per row of \code{x}.
}
\description{
Untibblify a data frame into a list of row lists
}
\keyword{internal}
