% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untibblify.R
\name{.untibblify_list}
\alias{.untibblify_list}
\title{Untibblify a named list using a spec}
\usage{
.untibblify_list(x, spec, call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{spec}{(\code{tspec} or \code{NULL}) A spec object describing the structure of
\code{x}.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A named list with fields converted according to \code{spec}.
}
\description{
Untibblify a named list using a spec
}
\keyword{internal}
