% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untibblify.R
\name{.untibblify_list_elt}
\alias{.untibblify_list_elt}
\title{Untibblify a single list element}
\usage{
.untibblify_list_elt(x, field_spec, call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{field_spec}{(\code{tib_collector}) A tibblify field collector.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The converted element, or \code{x} unchanged if no conversion applies.
}
\description{
Untibblify a single list element
}
\keyword{internal}
