% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untibblify.R
\name{.untibblify_row}
\alias{.untibblify_row}
\title{Untibblify a single data frame row into a named list}
\usage{
.untibblify_row(x, spec, call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{spec}{(\code{tspec} or \code{NULL}) A spec object describing the structure of
\code{x}.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
A named list with one element per field of \code{x}.
}
\description{
Untibblify a single data frame row into a named list
}
\keyword{internal}
