% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_object_utils.R
\name{.update_required_fields}
\alias{.update_required_fields}
\title{Update the required status of field specs}
\usage{
.update_required_fields(fields, required)
}
\arguments{
\item{fields}{(\code{list}) A named list of tib field specs.}

\item{required}{(\code{logical}) A named logical vector of required statuses, as
returned by \code{\link[=.get_required]{.get_required()}}.}
}
\value{
\code{fields} with the \verb{$required} component of each spec updated.
}
\description{
Update the required status of field specs
}
\keyword{internal}
