% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.vec_flatten}
\alias{.vec_flatten}
\title{Flatten a list to a vector}
\usage{
.vec_flatten(x, ptype, name_spec = zap())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{ptype}{(\code{vector(0)}) Prototype for the flattened output.}

\item{name_spec}{(\code{character(1)}, \code{function}, or \code{NULL}) Name specification
passed to \code{\link[vctrs:list_unchop]{vctrs::list_unchop()}}.}
}
\value{
A vector with elements from \code{x}.
}
\description{
Flatten a list to a vector
}
\keyword{internal}
