% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_df.R
\name{.vector_col_to_spec}
\alias{.vector_col_to_spec}
\title{Convert a vector column to a tib scalar or unspecified specification}
\usage{
.vector_col_to_spec(col, name)
}
\arguments{
\item{col}{(\code{any}) A column from a data frame, which may be a vector, a list,
or a nested data frame.}

\item{name}{(\code{character(1)}) The name of the field.}
}
\value{
A \code{\link[=tib_scalar]{tib_scalar()}} or \code{\link[=tib_unspecified]{tib_unspecified()}} specification.
}
\description{
Convert a vector column to a tib scalar or unspecified specification
}
\keyword{internal}
