% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.with_indexed_errors}
\alias{.with_indexed_errors}
\title{Wrap indexed purrr errors with context}
\usage{
.with_indexed_errors(
  expr,
  message,
  error_call = caller_env(),
  env = caller_env()
)
}
\arguments{
\item{expr}{(\code{any}) An expression to evaluate and return, with indexed errors
wrapped.}

\item{message}{(\code{character}) A cli message template.}

\item{error_call}{(\code{environment}) The environment to use for error messages.}

\item{env}{(\code{environment}) The environment used to evaluate glue fields in
\code{message}.}
}
\value{
The evaluated result of \code{expr}, or an error with added context.
}
\description{
Wrap indexed purrr errors with context
}
\keyword{internal}
