% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-tib.R, R/format-tibblify_object.R,
%   R/format-tspec.R
\name{formatting}
\alias{formatting}
\alias{print.tib_collector}
\alias{format.tib_scalar}
\alias{format.tib_variant}
\alias{format.tib_vector}
\alias{format.tib_unspecified}
\alias{format.tib_scalar_chr_date}
\alias{format.tib_vector_chr_date}
\alias{format.tib_row}
\alias{format.tib_df}
\alias{format.tib_recursive}
\alias{print.tibblify_object}
\alias{print.tspec}
\alias{format.tspec_df}
\alias{format.tspec_row}
\alias{format.tspec_recursive}
\alias{format.tspec_object}
\title{Printing tibblify specifications}
\usage{
\method{print}{tib_collector}(x, width = NULL, ..., names = NULL, fully_qualify = FALSE)

\method{format}{tib_scalar}(
  x,
  ...,
  .fill = NULL,
  .ptype_inner = NULL,
  .transform = NULL,
  multi_line = FALSE,
  nchar_indent = 0,
  width = NULL,
  names = FALSE,
  fully_qualify = FALSE
)

\method{format}{tib_variant}(
  x,
  ...,
  multi_line = FALSE,
  nchar_indent = 0,
  width = NULL,
  fully_qualify = FALSE
)

\method{format}{tib_vector}(
  x,
  ...,
  multi_line = FALSE,
  nchar_indent = 0,
  width = NULL,
  fully_qualify = FALSE
)

\method{format}{tib_unspecified}(
  x,
  ...,
  .fill = NULL,
  .ptype_inner = NULL,
  .transform = NULL,
  multi_line = FALSE,
  nchar_indent = 0,
  width = NULL,
  names = FALSE,
  fully_qualify = FALSE
)

\method{format}{tib_scalar_chr_date}(
  x,
  ...,
  multi_line = FALSE,
  nchar_indent = 0,
  width = NULL,
  fully_qualify = FALSE
)

\method{format}{tib_vector_chr_date}(
  x,
  ...,
  multi_line = FALSE,
  nchar_indent = 0,
  width = NULL,
  fully_qualify = FALSE
)

\method{format}{tib_row}(x, ..., width = NULL, names = NULL, fully_qualify = FALSE)

\method{format}{tib_df}(x, ..., width = NULL, names = NULL, fully_qualify = FALSE)

\method{format}{tib_recursive}(x, ..., width = NULL, names = NULL, fully_qualify = FALSE)

\method{print}{tibblify_object}(x, ...)

\method{print}{tspec}(x, width = NULL, ..., names = NULL, fully_qualify = FALSE)

\method{format}{tspec_df}(x, width = NULL, ..., names = NULL, fully_qualify = FALSE)

\method{format}{tspec_row}(x, width = NULL, ..., names = NULL, fully_qualify = FALSE)

\method{format}{tspec_recursive}(x, width = NULL, ..., names = NULL, fully_qualify = FALSE)

\method{format}{tspec_object}(x, width = NULL, ..., names = NULL, fully_qualify = FALSE)
}
\arguments{
\item{x}{(\code{any}) The spec to format or print.}

\item{width}{(\code{integer(1)}) The width (in number of characters) of text
output to generate.}

\item{...}{These dots are for future extensions and must be empty.}

\item{names}{(\code{logical(1)}) Should names be printed even if they can be
deduced from the spec?}

\item{fully_qualify}{(\code{logical(1)}) Should \verb{tib_*()} and \verb{tspec_*()} calls be
prefixed with \verb{tibblify::}?}

\item{.fill}{(\code{vector} or \code{NULL}) Optionally, a value to use if the field
is empty or does not exist. Note: this value must match the \code{.ptype_inner}
of the field (the value \emph{before} any transformation), not the \code{.ptype}.}

\item{.ptype_inner}{(\code{vector(0)}) A prototype of the input field.}

\item{.transform}{(\code{function} or \code{NULL}) A function to apply to the whole
vector after casting to \code{.ptype_inner}.}

\item{multi_line}{(\code{logical(1)}) Should the output be formatted across
multiple lines? For example, should each element of even a short list be
displayed on its own line?}

\item{nchar_indent}{(\code{integer(1)}) The number of (additional) characters that
will be used to indent the output when \code{multi_line = TRUE}. Primarily for
internal use when formatting is applied recursively.}
}
\value{
For \code{print()} methods, \code{x} is returned invisibly. \code{format()} methods
return a length-1 character vector.
}
\description{
The \code{print()} and \code{format()} methods for tibblify specifications provide the
code necessary to generate the specification. Function names are color-coded
to help visually distinguish different types of collectors.
}
\examples{
spec <- tspec_df(
  a = tib_int("a"),
  new_name = tib_chr("b"),
  row = tib_row(
    "row",
    x = tib_int("x")
  )
)
print(spec, names = FALSE)
print(spec, names = TRUE)
}
