% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibblify.R
\name{get_spec}
\alias{get_spec}
\title{Examine the column specification}
\usage{
get_spec(x)
}
\arguments{
\item{x}{(\code{data.frame}) The data frame to extract a spec from.}
}
\value{
A tibblify specification as returned by \code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}},
\code{\link[=tspec_object]{tspec_object()}}, or \code{\link[=tspec_recursive]{tspec_recursive()}}.
}
\description{
Examine the column specification
}
\examples{
df <- tibblify(list(list(x = 1, y = "a"), list(x = 2)))
get_spec(df)
}
