% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gh_repos}
\alias{gh_repos}
\title{GitHub Repositories}
\format{
A list of lists.
}
\usage{
gh_repos
}
\description{
A dataset containing some basic information about some GitHub repositories.
}
\keyword{datasets}
