% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gh_users}
\alias{gh_users}
\title{GitHub Users}
\format{
A list of lists.
}
\usage{
gh_users
}
\description{
A dataset containing some basic information about six GitHub users.
}
\keyword{datasets}
