% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{got_chars}
\alias{got_chars}
\title{Game of Thrones POV characters}
\format{
A unnamed list with 30 components, each representing a POV character.
Each character's component is a named list of length 18, containing
information such as name, aliases, and house allegiances.
}
\source{
\url{https://anapioficeandfire.com}
}
\usage{
got_chars
}
\description{
The data is from the \href{https://github.com/jennybc/repurrrsive}{repurrrsive package}.
}
\details{
Info on the point-of-view (POV) characters from the first five books in the
Song of Ice and Fire series by George R. R. Martin. Retrieved from An API Of
Ice And Fire.
}
\examples{
got_chars
str(lapply(got_chars, `[`, c("name", "culture")))
}
\concept{Game of Thrones data and functions}
\keyword{datasets}
