% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_shape.R, R/guess_spec.R
\name{guess_spec}
\alias{guess_spec}
\title{Guess the \code{tibblify()} Specification}
\usage{
guess_spec(x)

guess_spec(x)
}
\arguments{
\item{x}{A nested list.}
}
\value{
A specification object that can used in \code{tibblify()}.

A specification object that can used in \code{tibblify()}.
}
\description{
Guess the \code{tibblify()} Specification

Guess the \code{tibblify()} Specification
}
\examples{
guess_spec(list(x = 1, y = "a"))
guess_spec(list(list(x = 1), list(x = 2)))

guess_spec(gh_users)
guess_spec(list(x = 1, y = "a"))
guess_spec(list(list(x = 1), list(x = 2)))

guess_spec(gh_users)
}
