% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec.R, R/guess_tspec_df.R,
%   R/guess_tspec_list.R, R/guess_tspec_object.R, R/guess_tspec_object_list.R
\name{guess_tspec}
\alias{guess_tspec}
\alias{guess_tspec_df}
\alias{guess_tspec_list}
\alias{guess_tspec_object}
\alias{guess_tspec_object_list}
\title{Guess the \code{tibblify()} specification}
\usage{
guess_tspec(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  inform_unspecified = should_inform_unspecified(),
  call = rlang::caller_env()
)

guess_tspec_df(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  inform_unspecified = should_inform_unspecified(),
  call = rlang::current_call(),
  arg = rlang::caller_arg(x)
)

guess_tspec_list(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  inform_unspecified = should_inform_unspecified(),
  arg = caller_arg(x),
  call = current_call()
)

guess_tspec_object(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  inform_unspecified = should_inform_unspecified(),
  call = rlang::current_call()
)

guess_tspec_object_list(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  inform_unspecified = should_inform_unspecified(),
  arg = caller_arg(x),
  call = current_call()
)
}
\arguments{
\item{x}{(\code{list}) A nested list.}

\item{...}{These dots are for future extensions and must be empty.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{simplify_list}{(\code{logical(1)}) Should scalar lists be simplified to
vectors?}

\item{inform_unspecified}{(\code{logical(1)}) Inform about fields whose type could
not be determined?}

\item{call}{(\code{environment}) The environment to use for error messages.}

\item{arg}{(\code{character(1)}) An argument name. This name will be mentioned in
error messages as the input that is at the origin of a problem.}
}
\value{
A specification object that can be used in \code{\link[=tibblify]{tibblify()}}.
}
\description{
\code{guess_tspec()} automatically dispatches to the other \verb{guess_tspec_*()}
functions based on the shape of the input. If you are unhappy with its
output, calling a specific \verb{guess_tspec_*()} function may yield better
results, or at least clearer error messages about why that type isn't
supported.
\itemize{
\item Use \code{guess_tspec_df()} if the input is a data frame.
\item Use \code{guess_tspec_object()} if the input is an object (such as a JSON
object that has been read into R as a named list).
\item Use \code{guess_tspec_object_list()} if the input is a list of objects (such as
a JSON object that has been read into R as a list of named lists).
\item Use \code{guess_tspec_list()} if the input object is a list but you aren't sure
how it should be processed.
}

See \verb{vignette("supported-structures)} for a discussion of the input types
supported by tibblify.
}
\examples{
guess_tspec(list(x = 1, y = "a"))
guess_tspec(list(list(x = 1), list(x = 2)))

guess_tspec(gh_users)
}
