% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_specs.R
\name{lcol_lgl}
\alias{lcol_lgl}
\alias{lcol_int}
\alias{lcol_dbl}
\alias{lcol_chr}
\alias{lcol_dat}
\alias{lcol_dtt}
\alias{lcol_vec}
\alias{lcol_guess}
\alias{lcol_skip}
\alias{lcol_lst}
\alias{lcol_lst_of}
\alias{lcol_df_lst}
\alias{lcol_df}
\title{Create column specificcation}
\usage{
lcol_lgl(path, .default = zap(), .parser = NULL)

lcol_int(path, .default = zap(), .parser = NULL)

lcol_dbl(path, .default = zap(), .parser = NULL)

lcol_chr(path, .default = zap(), .parser = NULL)

lcol_dat(path, .default = zap(), .parser = NULL)

lcol_dtt(path, .default = zap(), .parser = NULL)

lcol_vec(path, ptype, .default = zap(), .parser = NULL)

lcol_guess(path, .default = NULL)

lcol_skip(path)

lcol_lst(path, .default = zap(), .parser = NULL)

lcol_lst_of(path, .ptype, .default = zap(), .parser = NULL)

lcol_df_lst(path, ..., .default = zap())

lcol_df(path, ..., .default = zap())
}
\arguments{
\item{path}{A character vector or list that is converted to an extractor function (similar to the \code{.f} argument in  \code{purrr::map()}).}

\item{.default}{Value to use if target is empty or absent. If \code{zap()} (the default) an error is thrown if target is empty or absent.}

\item{.parser}{A transformation applied to each element of the list before coercing to \code{.ptype}. This is usually needed for \code{lcol_dat()} and \code{lcol_dtt()}.}

\item{ptype}{The prototype of the vector.}

\item{.ptype}{The \code{.ptype} for \code{vctrs::list_of()}.}

\item{...}{Column specification passed on to \code{lcols()}.}
}
\description{
\code{lcols()} includes all fields in the input data, skipping the column types as the default.
}
