% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcols.R
\name{lcols}
\alias{lcols}
\title{Create list parser specification}
\usage{
lcols(..., .default = lcol_skip(zap()))
}
\arguments{
\item{...}{Column specification objects created by \verb{lcol_*()}.}

\item{.default}{The default parser to use.}
}
\description{
Create list parser specification
}
