% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listparser.R
\docType{package}
\name{listparser-package}
\alias{listparser}
\alias{listparser-package}
\title{listparser: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Maximilian Girlich \email{maximilian.girlich@metoda.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
