% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{parse_openapi_spec}
\alias{parse_openapi_spec}
\alias{parse_openapi_schema}
\title{Parse an OpenAPI spec}
\usage{
parse_openapi_spec(file)

parse_openapi_schema(file)
}
\arguments{
\item{file}{(\code{character(1)}) A path to a file, a connection, or literal data.}
}
\value{
For \code{parse_openapi_spec()}, a nested data frame with the columns
\itemize{
\item \code{endpoint} (\code{character}) Name of the endpoint.
\item \code{operations} (\code{list}) A list of data frames describing that endpoint.
See the \href{https://spec.openapis.org/oas/latest.html#paths-object}{Paths Object in the OpenAPI spec} for details.
All references (\verb{$ref}) in the spec are resolved.
}

For \code{parse_openapi_schema()}, a tibblify spec. All references (\verb{$ref}) in
the spec are resolved.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \href{https://www.openapis.org/}{OpenAPI Initiative} is a \href{https://www.linuxfoundation.org/projects}{Linux Foundation} project to define an
\href{https://spec.openapis.org/oas/latest.html}{OpenAPI Specification}, a formal
standard for describing HTTP APIs. Use \code{parse_openapi_spec()} to parse such
OpenAPI specs. You can also parse \href{https://spec.openapis.org/oas/latest.html#schema-object}{OpenAPI Schema Objects} (which
describe the structure of input and output datatypes) directly with
\code{parse_openapi_schema()}.
}
\examples{
file <- '{
  "$schema": "http://json-schema.org/draft-04/schema",
  "title": "Starship",
  "description": "A vehicle.",
  "type": "object",
  "properties": {
    "name": {
      "type": "string",
      "description": "The name of this vehicle. The common name, e.g. Sand Crawler."
    },
    "model": {
      "type": "string",
      "description": "The model or official name of this vehicle."
    },
    "url": {
      "type": "string",
      "format": "uri",
      "description": "The hypermedia URL of this resource."
    },
    "edited": {
      "type": "string",
      "format": "date-time",
      "description": "the ISO 8601 date format of the time this resource was edited."
    }
  },
  "required": [
    "name",
    "model",
    "edited"
  ]
}'
parse_openapi_schema(file)

# Spec example from https://swagger.io/docs/specification/v3_0/basic-structure/
spec_path <- system.file(
  "examples", "openapi", "sample_api.yaml", package = "tibblify"
)
spec <- parse_openapi_spec(spec_path)
spec
}
