% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-swagger.R
\name{parse_swagger_spec}
\alias{parse_swagger_spec}
\alias{parse_swagger_schema}
\title{Parse a swagger spec}
\usage{
parse_swagger_spec(file)

parse_swagger_schema(file)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (a
single string).}
}
\value{
For \code{parse_swagger_spec()} a data frame with the columns
\itemize{
\item \code{endpoint} \verb{<character>} Name of the endpoint.
\item \code{operation} \verb{<character>} The http operation; one of \code{"get"}, \code{"put"},
\code{"post"}, \code{"delete"}, \code{"options"}, \code{"head"}, \code{"patch"}, or \code{"trace"}.
\item \code{status_code} \verb{<character>} The http status code. May contain wildcards like
\verb{2xx} for all response codes between \code{200} and \code{299}.
\item \code{media_type} \verb{<character>} The media type.
\item \code{spec} \verb{<list>} A list of tibblify specifications.
}

For \code{parse_swagger_schema()} a tibblify spec.
}
\description{
Use \code{parse_swagger_spec()} to parse a \href{https://swagger.io/specification/}{Swagger spec}
or use \code{parse_swagger_schema()} to parse a Swagger schema.
}
\examples{
file <- '{
  "$schema": "http://json-schema.org/draft-04/schema",
  "title": "Starship",
  "description": "A vehicle.",
  "type": "object",
  "properties": {
    "name": {
      "type": "string",
      "description": "The name of this vehicle. The common name, such as Sand Crawler."
    },
    "model": {
      "type": "string",
      "description": "The model or official name of this vehicle. Such as All Terrain Attack Transport."
    },
    "url": {
      "type": "string",
      "format": "uri",
      "description": "The hypermedia URL of this resource."
    },
    "edited": {
      "type": "string",
      "format": "date-time",
      "description": "the ISO 8601 date format of the time that this resource was edited."
    }
  },
  "required": [
    "name",
    "model",
    "edited"
  ]
}'

parse_swagger_schema(file)
}
