% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{politicians}
\alias{politicians}
\title{Politicians}
\format{
A list of lists.
}
\usage{
politicians
}
\description{
A dataset containing some basic information about some politicians.
}
\keyword{datasets}
