% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec.R
\name{spec_df}
\alias{spec_df}
\alias{spec_object}
\alias{spec_row}
\title{Create a Tibblify Specification}
\usage{
spec_df(..., .names_to = NULL)

spec_object(...)

spec_row(...)
}
\description{
Create a Tibblify Specification
}
