% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_spec.R
\name{spec_guess}
\alias{spec_guess}
\alias{spec_guess_df}
\alias{spec_guess_list}
\alias{spec_guess_object_list}
\alias{spec_guess_object}
\title{Guess the \code{tibblify()} Specification}
\usage{
spec_guess(x)

spec_guess_df(x)

spec_guess_list(x, simplify_list = TRUE)

spec_guess_object_list(x, simplify_list = TRUE)

spec_guess_object(x, simplify_list = TRUE)
}
\arguments{
\item{x}{A nested list.}

\item{simplify_list}{Try to simplify lists if possible?}
}
\value{
A specification object that can used in \code{tibblify()}.
}
\description{
Guess the \code{tibblify()} Specification
}
\examples{
spec_guess(list(x = 1, y = "a"))
spec_guess(list(list(x = 1), list(x = 2)))

spec_guess(gh_users)
}
