% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R, R/tib_spec_class.R,
%   R/tib_spec_other.R
\name{tib_spec}
\alias{tib_spec}
\alias{tib_scalar}
\alias{tib_vector}
\alias{tib_unspecified}
\alias{tib_lgl}
\alias{tib_int}
\alias{tib_dbl}
\alias{tib_chr}
\alias{tib_date}
\alias{tib_chr_date}
\alias{tib_lgl_vec}
\alias{tib_int_vec}
\alias{tib_dbl_vec}
\alias{tib_chr_vec}
\alias{tib_date_vec}
\alias{tib_chr_date_vec}
\alias{tib_variant}
\alias{tib_recursive}
\alias{tib_row}
\alias{tib_df}
\title{Create a field specification}
\usage{
tib_scalar(
  .key,
  .ptype,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = .ptype,
  .transform = NULL,
  key = deprecated(),
  ptype = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated()
)

tib_vector(
  .key,
  .ptype,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = .ptype,
  .transform = NULL,
  .elt_transform = NULL,
  .input_form = c("vector", "scalar_list", "object"),
  .values_to = NULL,
  .names_to = NULL,
  key = deprecated(),
  ptype = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated(),
  elt_transform = deprecated(),
  input_form = deprecated(),
  values_to = deprecated(),
  names_to = deprecated()
)

tib_unspecified(
  .key,
  ...,
  .required = TRUE,
  key = deprecated(),
  required = deprecated()
)

tib_lgl(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = logical(),
  .transform = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated()
)

tib_int(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = integer(),
  .transform = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated()
)

tib_dbl(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = double(),
  .transform = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated()
)

tib_chr(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = character(),
  .transform = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated()
)

tib_date(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = vctrs::new_date(),
  .transform = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated()
)

tib_chr_date(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .format = "\%Y-\%m-\%d",
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  format = deprecated()
)

tib_lgl_vec(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = logical(),
  .transform = NULL,
  .elt_transform = NULL,
  .input_form = c("vector", "scalar_list", "object"),
  .values_to = NULL,
  .names_to = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated(),
  elt_transform = deprecated(),
  input_form = deprecated(),
  values_to = deprecated(),
  names_to = deprecated()
)

tib_int_vec(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = integer(),
  .transform = NULL,
  .elt_transform = NULL,
  .input_form = c("vector", "scalar_list", "object"),
  .values_to = NULL,
  .names_to = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated(),
  elt_transform = deprecated(),
  input_form = deprecated(),
  values_to = deprecated(),
  names_to = deprecated()
)

tib_dbl_vec(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = double(),
  .transform = NULL,
  .elt_transform = NULL,
  .input_form = c("vector", "scalar_list", "object"),
  .values_to = NULL,
  .names_to = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated(),
  elt_transform = deprecated(),
  input_form = deprecated(),
  values_to = deprecated(),
  names_to = deprecated()
)

tib_chr_vec(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = character(),
  .transform = NULL,
  .elt_transform = NULL,
  .input_form = c("vector", "scalar_list", "object"),
  .values_to = NULL,
  .names_to = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated(),
  elt_transform = deprecated(),
  input_form = deprecated(),
  values_to = deprecated(),
  names_to = deprecated()
)

tib_date_vec(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = vctrs::new_date(),
  .transform = NULL,
  .elt_transform = NULL,
  .input_form = c("vector", "scalar_list", "object"),
  .values_to = NULL,
  .names_to = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  ptype_inner = deprecated(),
  transform = deprecated(),
  elt_transform = deprecated(),
  input_form = deprecated(),
  values_to = deprecated(),
  names_to = deprecated()
)

tib_chr_date_vec(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .input_form = c("vector", "scalar_list", "object"),
  .values_to = NULL,
  .names_to = NULL,
  .format = "\%Y-\%m-\%d",
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  input_form = deprecated(),
  values_to = deprecated(),
  names_to = deprecated(),
  format = deprecated()
)

tib_variant(
  .key,
  ...,
  .required = TRUE,
  .fill = NULL,
  .transform = NULL,
  .elt_transform = NULL,
  key = deprecated(),
  required = deprecated(),
  fill = deprecated(),
  transform = deprecated(),
  elt_transform = deprecated()
)

tib_recursive(.key, ..., .children, .children_to = .children, .required = TRUE)

tib_row(.key, ..., .required = TRUE)

tib_df(.key, ..., .required = TRUE, .names_to = NULL)
}
\arguments{
\item{.key, key}{(\code{character}) The path of names to the field in the object.}

\item{.ptype, ptype}{(\code{vector(0)}) A prototype of the desired output type of the
field.}

\item{...}{These dots are for future extensions and must be empty.}

\item{.required, required}{(\code{logical(1)}) Throw an error if the field does not exist?}

\item{.fill, fill}{(\code{vector} or \code{NULL}) Optionally, a value to use if the field
is empty or does not exist. Note: this value must match the \code{.ptype_inner}
of the field (the value \emph{before} any transformation), not the \code{.ptype}.}

\item{.ptype_inner, ptype_inner}{(\code{vector(0)}) A prototype of the input field.}

\item{.transform, transform}{(\code{function} or \code{NULL}) A function to apply to the whole
vector after casting to \code{.ptype_inner}.}

\item{.elt_transform, elt_transform}{(\code{function} or \code{NULL}) A function to apply to each
element before casting to \code{.ptype_inner}.}

\item{.input_form, input_form}{(\code{character(1)}) The structure of the input field. Can be
one of:
\itemize{
\item \code{"vector"}: The field is a vector, e.g. \code{c(1, 2, 3)}.
\item \code{"scalar_list"}: The field is a list of scalars, e.g. \code{list(1, 2, 3)}.
\item \code{"object"}: The field is a named list of scalars, e.g.
\code{list(a = 1, b = 2, c = 3)}.
}}

\item{.values_to, values_to}{(\code{character(1)} or \code{NULL}) For \code{NULL} (the default), the
field is converted to a \code{.ptype} vector. If a string is provided, the field
is converted to a tibble and the values go into the specified column.}

\item{.names_to, names_to}{(\code{character(1)} or \code{NULL}) What to do with the inner names
of the object. Can be one of:
\itemize{
\item \code{NULL}: the default. The inner names of the field are not used.
\item A string: Use only if the input form is \code{"object"} or \code{"vector"}, and
\code{.values_to} is a string. The inner names of the field will populate the
specified column in the field's tibble.
}}

\item{.format, format}{(\code{character(1)} or \code{NULL}) Passed to the \code{format} argument of
\code{\link[=as.Date]{as.Date()}}.}

\item{.children}{(\code{character(1)}) The name of the field that contains the
children.}

\item{.children_to}{(\code{character(1)}) The column name in which to store the
children.}
}
\value{
A tibblify field collector. This specification can be used with
\code{\link[=tspec_df]{tspec_df()}} or another \verb{tspec_*()} function to specify how to process an
object.
}
\description{
Use the \verb{tib_*()} functions to specify how to process the fields of an
object.
}
\details{
There are five families of \verb{tib_*()} functions:
\itemize{
\item \code{tib_scalar(.ptype)}: Cast each instance of the field to a length-one
vector of type \code{.ptype}. Inside \code{\link[=tspec_df]{tspec_df()}}, this results in a column of
the specified \code{.ptype}.
\item \code{tib_vector(.ptype)}: Cast each instance of the field to an arbitrary
length vector of type \code{.ptype}. Inside \code{\link[=tspec_df]{tspec_df()}}, this results in a list
column of vectors of the specified \code{.ptype}.
\item \code{tib_variant()}: Cast each instance of the field to a list. Inside
\code{\link[=tspec_df]{tspec_df()}}, this results in a list column of lists.
\item \code{tib_row()}: Cast each instance of the field to a 1-row tibble. Inside
\code{\link[=tspec_df]{tspec_df()}}, this results in a list column of 1-row tibbles.
\item \code{tib_df()}: Cast each instance of the field to a tibble. Inside
\code{\link[=tspec_df]{tspec_df()}}, this results in a list column of tibbles (each of which can
have multiple rows).

There are some special shortcuts of \code{tib_scalar()} and \code{tib_vector()} for
the most common prototypes:
\item \code{logical()}: \code{tib_lgl()} and \code{tib_lgl_vec()}
\item \code{integer()}: \code{tib_int()} and \code{tib_int_vec()}
\item \code{double()}: \code{tib_dbl()} and \code{tib_dbl_vec()}
\item \code{character()}: \code{tib_chr()} and \code{tib_chr_vec()}
\item \code{Date}: \code{tib_date()} and \code{tib_date_vec()}

Further, there are special shortcuts for dates encoded as character:
\code{tib_chr_date()} and \code{tib_chr_date_vec()}.

There are two other \verb{tib_*()} functions for special cases:
\item \code{tib_recursive()}: Cast each instance of the field to a tibble, within
which columns can themselves contain the same sorts of tibble, etc. Inside
\code{\link[=tspec_df]{tspec_df()}}, this results in a list column of tibbles, each row of which
can itself contain a tibble, etc. This is intended for structures such as a
directory tree.
\item \code{tib_unspecified()}: Tag a field in the object as unspecified. The
\code{unspecified} argument of \code{\link[=tibblify]{tibblify()}} controls how such fields are
handled. If you are constructing a specification manually (as opposed to
using \code{\link[=guess_tspec]{guess_tspec()}}), you should most likely specify such columns with
\code{tib_variant()}, or leave them out of the spec entirely.
}
}
\examples{
tib_int("int")
tib_int("int", .required = FALSE, .fill = 0)

# This is essentially how `tib_chr_date()` is implemented.
tib_scalar(
  "date",
  Sys.Date(),
  .transform = function(x) as.Date(x, format = "\%Y-\%m-\%d")
)

tib_df(
  "data",
  .names_to = "id",
  age = tib_int("age"),
  name = tib_chr("name")
)
}
