% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibblify-package.R
\docType{package}
\name{tibblify-package}
\alias{tibblify-package}
\title{tibblify: Rectangle Nested Lists}
\description{
A tool to rectangle a nested list, that is to convert it into a tibble. This is done automatically or according to a given specification. A common use case is for nested lists coming from parsing 'JSON' files, or the 'JSON' responses of 'REST' 'APIs'. Rectangling uses the 'vctrs' package, and therefore offers a wide support of vector types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tibblify.wrangle.zone}
  \item \url{https://github.com/wranglezone/tibblify}
  \item Report bugs at \url{https://github.com/wranglezone/tibblify/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Maximilian Girlich \email{maximilian.girlich@outlook.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Kirill Müller [contributor]
}

}
\keyword{internal}
