% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibblify.R
\name{tibblify}
\alias{tibblify}
\title{Rectangle a nested list}
\usage{
tibblify(x, spec = NULL, unspecified = NULL)
}
\arguments{
\item{x}{(\code{list}) A nested list.}

\item{spec}{(\code{tspec}) A specification of how to convert \code{x}. Generated with
\code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}}, \code{\link[=tspec_object]{tspec_object()}}, \code{\link[=tspec_recursive]{tspec_recursive()}}, or
\code{\link[=guess_tspec]{guess_tspec()}}. If \code{spec} is \code{NULL} (the default), \code{guess_tspec(x, inform_unspecified = TRUE)} will be used to guess the \code{spec}.}

\item{unspecified}{(\code{character(1)}) What to do with \code{\link[=tib_unspecified]{tib_unspecified()}}
fields. Can be one of
\itemize{
\item \code{"error"}: Throw an error.
\item \code{"inform"}: Inform the user then parse as with \code{\link[=tib_variant]{tib_variant()}}.
\item \code{"drop"}: Do not parse these fields.
\item \code{"list"}: Parse unspecified fields into lists as with \code{\link[=tib_variant]{tib_variant()}}.
}}
}
\value{
Either a tibble or a list, depending on the specification.
}
\description{
Transform a nested list into a tibble or a list of objects according to a
specification.
}
\details{
Fields specifically tagged as \code{\link[=tib_unspecified]{tib_unspecified()}} in the \code{spec} (or guessed
as such) will be handled according to the \code{unspecified} argument. Fields that
are present in \code{x} but not mentioned in the \code{spec} are ignored.
}
\examples{
# List of Objects -----------------------------------------------------------
x <- list(
  list(id = 1, name = "Tyrion Lannister"),
  list(id = 2, name = "Victarion Greyjoy")
)
tibblify(x)

# Provide a specification
spec <- tspec_df(
  id = tib_int("id"),
  name = tib_chr("name")
)
tibblify(x, spec)

# Object --------------------------------------------------------------------
# Provide a specification for a single object
tibblify(x[[1]], tspec_object(spec))

# Recursive Trees -----------------------------------------------------------
x <- list(
  list(
    id = 1,
    name = "a",
    children = list(
      list(id = 11, name = "aa"),
      list(id = 12, name = "ab", children = list(
        list(id = 121, name = "aba")
      ))
    ))
)
spec <- tspec_recursive(
  tib_int("id"),
  tib_chr("name"),
  .children = "children"
)
out <- tibblify(x, spec)
out
out$children
out$children[[1]]$children[[2]]
}
\seealso{
Use \code{\link[=untibblify]{untibblify()}} to undo the result of \code{tibblify()}.
}
