% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_combine.R
\name{tspec_combine}
\alias{tspec_combine}
\title{Combine multiple specifications}
\usage{
tspec_combine(...)
}
\arguments{
\item{...}{(\code{tspec}) Specifications to combine.}
}
\value{
A tibblify specification.
}
\description{
Combine specifications created by \code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}}, or
\code{\link[=tspec_object]{tspec_object()}}. The resulting specification includes all fields from the
input specifications.
}
\details{
If a field is specified in multiple input specifications, the field
specifications will be combined to produce a single field specification,
using the most specific specification for each argument. See the examples for
details.
}
\examples{
# union of fields
tspec_combine(
  tspec_df(tib_int("a")),
  tspec_df(tib_chr("b"))
)

# unspecified + x -> x
tspec_combine(
  tspec_df(tib_unspecified("a")),
  tspec_df(tib_int("a"))
)

# scalar + vector -> vector
tspec_combine(
  tspec_df(tib_chr("a")),
  tspec_df(tib_chr_vec("a"))
)

# scalar/vector + variant -> variant
tspec_combine(
  tspec_df(tib_chr("a")),
  tspec_df(tib_chr_vec("a")),
  tspec_df(tib_variant("a"))
)
}
