% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspec.R
\name{tspec_df}
\alias{tspec_df}
\alias{tspec_object}
\alias{tspec_row}
\alias{tspec_recursive}
\title{Create a tibblify specification}
\usage{
tspec_df(
  ...,
  .input_form = c("rowmajor", "colmajor"),
  .names_to = NULL,
  .vector_allows_empty_list = FALSE,
  vector_allows_empty_list = deprecated()
)

tspec_object(
  ...,
  .input_form = c("rowmajor", "colmajor"),
  .vector_allows_empty_list = FALSE,
  vector_allows_empty_list = deprecated()
)

tspec_row(
  ...,
  .input_form = c("rowmajor", "colmajor"),
  .vector_allows_empty_list = FALSE,
  vector_allows_empty_list = deprecated()
)

tspec_recursive(
  ...,
  .children,
  .children_to = .children,
  .input_form = c("rowmajor", "colmajor"),
  .vector_allows_empty_list = FALSE,
  vector_allows_empty_list = deprecated()
)
}
\arguments{
\item{...}{(\code{tib_collector} or \code{tspec}) Column specifications created by
\verb{tib_*()} or \verb{tspec_*()}. If the dots are named, the name will be used for
the resulting column. Otherwise, the name of the input will be used for the
column name.}

\item{.input_form}{(\code{character(1)}) The input form of data-frame-like lists.
Can be one of:
\itemize{
\item \code{"rowmajor"}: The default. The input is a named list of rows.
\item \code{"colmajor"}: The input is a named list of columns.
}}

\item{.names_to}{(\code{character(1)} or \code{NULL}) The name of the column in the
output which will contain the names of top-level elements of the input
named list. If \code{NULL}, the default, no name column is created.}

\item{.vector_allows_empty_list, vector_allows_empty_list}{(\code{logical(1)})
Should empty lists for columns with \code{.input_form = "vector"} be accepted
and treated as empty vectors?}

\item{.children}{(\code{character(1)}) The name of the field that contains the
children.}

\item{.children_to}{(\code{character(1)}) The column name in which to store the
children.}
}
\value{
A tibblify specification.
}
\description{
Use \code{tspec_df()} to specify how to convert a list of objects to a tibble. Use
\code{tspec_row()} to specify how to convert an object to a one-row tibble. Use
\code{tspec_object()} to specify how to convert an object to a list.
}
\details{
In column-major format, all fields are required, regardless of the
\code{.required} argument.
}
\examples{
tspec_df(
  id = tib_int("id"),
  name = tib_chr("name"),
  aliases = tib_chr_vec("aliases")
)

# Equivalent to
tspec_df(
  tib_int("id"),
  tib_chr("name"),
  tib_chr_vec("aliases")
)

# To create multiple columns of the same type use the bang-bang-bang (`!!!`)
# operator together with `purrr::map()`
tspec_df(
  !!!purrr::map(purrr::set_names(c("id", "age")), tib_int),
  !!!purrr::map(purrr::set_names(c("name", "title")), tib_chr)
)

# The `tspec_*()` functions can also be nested
spec1 <- tspec_object(
  int = tib_int("int"),
  chr = tib_chr("chr")
)
spec2 <- tspec_object(
  int2 = tib_int("int2"),
  chr2 = tib_chr("chr2")
)

tspec_df(spec1, spec2)
}
