% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tree.R
\name{unnest_tree}
\alias{unnest_tree}
\title{Unnest a recursive data frame}
\usage{
unnest_tree(
  x,
  id_col,
  child_col,
  level_to = "level",
  parent_to = "parent",
  ancestors_to = NULL
)
}
\arguments{
\item{x}{(\code{data.frame}) The data frame to unnest.}

\item{id_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
uniquely identifies each observation.}

\item{child_col}{(\code{character(1)}, \code{integer(1)}, or \code{symbol}) The column that
contains the children of an observation. This column must be a list where
each element is either \code{NULL} or a data frame with the same columns as \code{x}.}

\item{level_to}{(\code{character(1)}) The column name (\code{"level"} by default) in
which to store the level of an observation. Use \code{NULL} if you don't need
this information.}

\item{parent_to}{(\code{character(1)}) The column name (\code{"parent"} by default) in
which to store the parent id of an observation. Use \code{NULL} if you don't
need this information.}

\item{ancestors_to}{(\code{character(1)}) The column name (\code{NULL} by default) in
which to store the ids of the ancestors of a deeply nested child. Use
\code{NULL} if you don't need this information.}
}
\value{
A "flat" data frame.
}
\description{
Unnest a recursive data frame
}
\examples{
df <- tibble(
  id = 1L,
  name = "a",
  children = list(
    tibble(
      id = 11:12,
      name = c("b", "c"),
      children = list(
        NULL,
        tibble(
          id = 121:122,
          name = c("d", "e")
        )
      )
    )
  )
)
df

unnest_tree(
  df,
  id_col = "id",
  child_col = "children",
  level_to = "level",
  parent_to = "parent",
  ancestors_to = "ancestors"
)
}
