% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untibblify.R
\name{untibblify}
\alias{untibblify}
\title{Convert a data frame or object into a nested list}
\usage{
untibblify(x, spec = get_spec(x))
}
\arguments{
\item{x}{(\code{data.frame} or \code{object}) An object to convert into a nested list.}

\item{spec}{(\code{tspec}) Optional. A spec object which was used to create \code{x}.
Defaults to the spec stored as the \code{tib_spec} attribute of \code{x}, if present.}
}
\value{
A nested list.
}
\description{
Convert a data frame or an object into a nested list. This is the inverse
operation of \code{tibblify()}. See \code{vignette("supported-structures")} for a
description of objects recognized by tibblify.
}
\examples{
x <- tibble(
  a = 1:2,
  b = tibble(
    x = c("a", "b"),
    y = c(1.5, 2.5)
  )
)
untibblify(x)
}
