% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.cast_fill}
\alias{.cast_fill}
\title{Cast a fill value using stbl-style friendlier coercion}
\usage{
.cast_fill(fill, ptype, ptype_arg = NULL, call = caller_env())
}
\arguments{
\item{fill}{The fill value to coerce.}

\item{ptype}{The target prototype.}

\item{ptype_arg}{Argument name for the ptype, used in error messages.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The coerced fill value.
}
\description{
Cast a fill value using stbl-style friendlier coercion
}
\keyword{internal}
