% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_tspec_utils.R
\name{.cast_posixlt_ptype}
\alias{.cast_posixlt_ptype}
\title{Conver POSIXlt to POSIXct}
\usage{
.cast_posixlt_ptype(ptype)
}
\arguments{
\item{ptype}{(\code{any}) An object which might be a \code{POSIXlt} or have \code{POSIXlt}
in its class hierarchy.}
}
\value{
An object of class \code{POSIXct} if the input is \code{POSIXlt} (to be in
line with \url{https://github.com/r-lib/vctrs/issues/1576}), otherwise the
input unchanged.
}
\description{
Conver POSIXlt to POSIXct
}
\keyword{internal}
