% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_arg_different}
\alias{.check_arg_different}
\title{Validate that arguments are different}
\usage{
.check_arg_different(arg, ..., arg_name = caller_arg(arg), call = caller_env())
}
\arguments{
\item{arg}{(\code{any}) The value to compare against \code{...}.}

\item{...}{Other arguments that \code{arg} must differ from.}

\item{arg_name}{(\code{character(1)}) The argument name shown in error messages.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
\code{NULL} (invisibly). Throws an error if any values are identical.
}
\description{
Validate that arguments are different
}
\keyword{internal}
