% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tree.R
\name{.check_col_value_duplicates}
\alias{.check_col_value_duplicates}
\title{Check that a column has no duplicate values}
\usage{
.check_col_value_duplicates(x, x_arg, call = caller_env())
}
\arguments{
\item{x}{(\code{any}) The object to check.}

\item{x_arg}{(\code{character(1)}) The name of the x argument. This name will be
mentioned in error messages as the input that is at the origin of a
problem.}

\item{call}{(\code{environment}) The environment to use for error messages.}
}
\value{
The input \code{x}.
}
\description{
Check that a column has no duplicate values
}
\keyword{internal}
