% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.collect_parts}
\alias{.collect_parts}
\title{Collect formatted field parts}
\usage{
.collect_parts(fields, width, force_names, args)
}
\arguments{
\item{fields}{(\code{list}) Fields to format.}

\item{width}{(\code{integer(1)}) The width (in number of characers) of text output
to generate.}

\item{force_names}{(\code{logical(1)}) Should names be printed even if they can be
deduced from the spec?}

\item{args}{(\code{list}) Additional arguments to format and prepend before
fields.}
}
\value{
A character vector of formatted field parts.
}
\description{
Collect formatted field parts
}
\keyword{internal}
