% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_tspec_object.R
\name{.finalize_tspec_object}
\alias{.finalize_tspec_object}
\title{Finalize a tibblify object}
\usage{
.finalize_tspec_object(field_spec, field)
}
\arguments{
\item{field_spec}{(\code{tib_collector}) A field specification.}

\item{field}{(\code{any}) The field value.}
}
\description{
Finalize a tibblify object
}
\keyword{internal}
