% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-utils.R
\name{.format_field_canonical_names}
\alias{.format_field_canonical_names}
\title{Format field parts with canonical names}
\usage{
.format_field_canonical_names(fields, width, force_names, fully_qualify)
}
\arguments{
\item{fields}{(\code{list} or \code{NULL}) Fields to format.}

\item{width}{(\code{integer(1)}) The width (in number of characters) of text
output to generate.}

\item{force_names}{(\code{logical(1)}) Should names be printed even if they can be
deduced from the spec?}

\item{fully_qualify}{(\code{logical(1)}) Should printed/formatted \verb{tib_*()} and
\verb{tspec_*()} calls be prefixed with \verb{tibblify::}?}
}
\value{
A character vector of formatted field parts with canonical names
potentially suppressed.
}
\description{
Format field parts with canonical names
}
\keyword{internal}
