% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-tib.R
\name{.format_fill}
\alias{.format_fill}
\title{Format a fill value as a character string}
\usage{
.format_fill(x)
}
\arguments{
\item{x}{A fill value.}
}
\value{
(\code{character(1)}) An R expression string representing \code{x}.
}
\description{
Format a fill value as a character string
}
\keyword{internal}
