% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-tib.R
\name{.format_ptype}
\alias{.format_ptype}
\title{Format a ptype object as a character string}
\usage{
.format_ptype(x)
}
\arguments{
\item{x}{(\code{ptype}) A prototype object.}
}
\value{
(\code{character(1)}) An R expression string representing \code{x}.
}
\description{
Format a ptype object as a character string
}
\keyword{internal}
