% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-tib.R
\name{.format_tib_f}
\alias{.format_tib_f}
\title{Return the display function name for a collector}
\usage{
.format_tib_f(x)
}
\arguments{
\item{x}{A tibblify collector object.}
}
\value{
(\code{character(1)}) The (possibly ANSI-colored) function name string.
}
\description{
Return the display function name for a collector
}
\keyword{internal}
