% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_to_tspec.R
\name{.get_field_of_type}
\alias{.get_field_of_type}
\title{Extract a field of a specific type from a spec}
\usage{
.get_field_of_type(spec, name, type, fn_name, .call = caller_env())
}
\arguments{
\item{spec}{(\code{tspec} or \code{NULL}) A spec object describing the structure of
\code{x}.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{type}{(\code{character(1)}) The expected field type string.}

\item{fn_name}{(\code{character(1)}) The name of the tib constructor function
for use in error messages.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{tib_collector}) The field spec.
}
\description{
Extract a field of a specific type from a spec
}
\keyword{internal}
