% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_inform_unspecified.R
\name{.get_unspecified_paths}
\alias{.get_unspecified_paths}
\title{Find unspecified fields}
\usage{
.get_unspecified_paths(spec)
}
\arguments{
\item{spec}{(\code{tspec}) A specification of how to convert \code{x}. Generated with
\code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}}, \code{\link[=tspec_object]{tspec_object()}}, \code{\link[=tspec_recursive]{tspec_recursive()}}, or (if
\code{NULL}, the default), \code{\link[=guess_tspec]{guess_tspec()}}.}
}
\value{
A list of \code{\link[=tib_unspecified]{tib_unspecified()}} objects and objects containing
\code{\link[=tib_unspecified]{tib_unspecified()}} objects.
}
\description{
Find unspecified fields
}
\keyword{internal}
