% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.handle_one_of_tspec}
\alias{.handle_one_of_tspec}
\title{Handle a \code{oneOf} schema by combining sub-schemas into a tspec}
\usage{
.handle_one_of_tspec(schema, openapi_spec)
}
\arguments{
\item{schema}{(\code{list}) A JSON schema object, as defined in the
\href{https://spec.openapis.org/oas/v3.1.0#schema-object}{Schema Object},
typically from \code{openapi_spec$components$schemas} or inline within the
spec.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A combined tspec, or \code{tib_variant("dummy")} if schemas are
incompatible.
}
\description{
Handle a \code{oneOf} schema by combining sub-schemas into a tspec
}
\keyword{internal}
