% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.openapi_resolve_reference}
\alias{.openapi_resolve_reference}
\title{Resolve \verb{$ref} references in an OpenAPI schema object}
\usage{
.openapi_resolve_reference(schema, openapi_spec)
}
\arguments{
\item{schema}{(\code{list}) A JSON schema object, typically from
\code{openapi_spec$components$schemas} or inline within the spec.}

\item{openapi_spec}{(\code{list}) The parsed OpenAPI specification.}
}
\value{
(\code{list}) The schema with all \verb{$ref} references resolved.
}
\description{
Resolve \verb{$ref} references in an OpenAPI schema object
}
\keyword{internal}
