% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_parameters}
\alias{.parse_parameters}
\title{Parse a list of parameter objects from an OpenAPI spec}
\usage{
.parse_parameters(parameters, openapi_spec)
}
\arguments{
\item{parameters}{(\code{list} or \code{NULL}) A list of parameter objects from an
OpenAPI spec, as defined in the \href{https://spec.openapis.org/oas/v3.1.0#parameter-object}{Parameter Object}.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
(\code{tbl_df} or \code{NULL}) A tibble of parsed parameters, or \code{NULL} if
\code{parameters} is \code{NULL}.
}
\description{
Parse a list of parameter objects from an OpenAPI spec
}
\keyword{internal}
