% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_response_object}
\alias{.parse_response_object}
\title{Parse a single response object from an OpenAPI spec}
\usage{
.parse_response_object(response_object, openapi_spec)
}
\arguments{
\item{response_object}{(\code{list}) A response object from an OpenAPI spec, as
defined in the
\href{https://spec.openapis.org/oas/v3.1.0#response-object}{Response Object}.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
(\code{tbl_df}) A one-row tibble describing the response.
}
\description{
Parse a single response object from an OpenAPI spec
}
\keyword{internal}
