% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_responses_object}
\alias{.parse_responses_object}
\title{Parse a responses object from an OpenAPI spec}
\usage{
.parse_responses_object(responses_object, openapi_spec)
}
\arguments{
\item{responses_object}{(\code{list}) A responses object from an OpenAPI spec,
mapping status codes to response objects, as defined in the \href{https://spec.openapis.org/oas/v3.1.0#responses-object}{Responses Object}.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
(\code{tbl_df}) A tibble of parsed response objects with a \code{status_code}
column.
}
\description{
Parse a responses object from an OpenAPI spec
}
\keyword{internal}
