% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_openapi.R
\name{.parse_schema}
\alias{.parse_schema}
\title{Parse an OpenAPI schema object into a tib field spec}
\usage{
.parse_schema(schema, name, openapi_spec)
}
\arguments{
\item{schema}{(\code{list}) A JSON schema object, as defined in the
\href{https://spec.openapis.org/oas/v3.1.0#schema-object}{Schema Object},
typically from \code{openapi_spec$components$schemas} or inline within the
spec.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{openapi_spec}{(\code{list}) A parsed OpenAPI specification.}
}
\value{
A tib field spec corresponding to the schema type.
}
\description{
Parse an OpenAPI schema object into a tib field spec
}
\keyword{internal}
