% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared_params.R
\name{.shared-params}
\alias{.shared-params}
\title{Shared parameters}
\arguments{
\item{arg}{(\code{character(1)}) An argument name. This name will be mentioned
in error messages as the input that is at the origin of a problem.}

\item{.call}{(\code{environment}) The environment to use for error messages.}

\item{.children}{(\code{character(1)}) The name of the field that contains the
children.}

\item{.children_to}{(\code{character(1)}) The column name in which to store the
children.}

\item{.elt_transform}{(\code{function} or \code{NULL}) A function to apply to each
element before casting to \code{.ptype_inner}.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{.fill}{(\code{vector} or \code{NULL}) Optionally, a value to use if the field
does not exist.}

\item{.format}{(\code{character(1)} or \code{NULL}) Passed to the \code{format} argument of
\code{\link[=as.Date]{as.Date()}}.}

\item{inform_unspecified}{(\code{logical(1)}) Inform about fields whose type could
not be determined?}

\item{.key}{(\code{character}) The path of names to the field in the object.}

\item{name}{(\code{character(1)}) The name of the field.}

\item{.ptype}{(\code{vector(0)}) A prototype of the desired output type of the
field.}

\item{.ptype_inner}{(\code{vector(0)}) A prototype of the input field.}

\item{.required}{(\code{logical(1)}) Throw an error if the field does not exist?}

\item{simplify_list}{(\code{logical(1)}) Should scalar lists be simplified to
vectors?}

\item{spec_list}{(\code{list}) A list of specifications.}

\item{tib_list}{(\code{list}) A list of tib fields.}

\item{.transform}{(\code{function} or \code{NULL}) A function to apply to the whole
vector after casting to \code{.ptype_inner}.}

\item{.values_to}{(\code{character(1)} or \code{NULL}) For \code{NULL} (the default), the
field is converted to a \code{.ptype} vector. If a string is provided, the field
is converted to a tibble and the values go into the specified column.}

\item{x}{(\code{any}) The object to check.}
}
\description{
These parameters are used in multiple functions. They are defined here to
make them easier to import and to find.
}
\keyword{internal}
