% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared_params.R
\name{.shared-params}
\alias{.shared-params}
\title{Shared parameters}
\arguments{
\item{allow_null}{(\code{logical(1)}) Whether \code{NULL} is accepted.}

\item{arg}{(\code{character(1)}) An argument name. This name will be mentioned in
error messages as the input that is at the origin of a problem.}

\item{.call}{(\code{environment}) The environment to use for error messages.}

\item{.children}{(\code{character(1)}) The name of the field that contains the
children.}

\item{.children_to}{(\code{character(1)}) The column name in which to store the
children.}

\item{col}{(\code{any}) A column from a data frame, which may be a vector, a list,
or a nested data frame.}

\item{elt}{(\code{character} or \code{integer}) An element of a path.}

\item{.elt_transform}{(\code{function} or \code{NULL}) A function to apply to each
element before casting to \code{.ptype_inner}.}

\item{empty_list_unspecified}{(\code{logical(1)}) Treat empty lists as
unspecified?}

\item{env}{(\code{environment}) The environment used to evaluate glue fields in
\code{message}.}

\item{.error_call}{(\code{environment}) The environment to use for error messages.}

\item{expr}{(\code{any}) An expression to evaluate and return, with indexed errors
wrapped.}

\item{field_spec}{(\code{tib_collector}) A tibblify field collector.}

\item{.fill}{(\code{vector} or \code{NULL}) Optionally, a value to use if the field
does not exist.}

\item{.format}{(\code{character(1)} or \code{NULL}) Passed to the \code{format} argument of
\code{\link[=as.Date]{as.Date()}}.}

\item{index}{(\code{integer(1)}) A zero-based location in a path.}

\item{inform_unspecified}{(\code{logical(1)}) Inform about fields whose type could
not be determined?}

\item{input_form}{(\code{character(1)}) The input form string used in error
messages.}

\item{.key}{(\code{character}) The path of names to the field in the object.}

\item{local_env}{(\code{environment}) A local environment used to track state
across recursive calls, such as whether empty lists were encountered.}

\item{message}{(\code{character}) A cli message template.}

\item{multi_line}{(\code{logical(1)}) Should the output be formatted across
multiple lines? For example, should each element of even a short list be
displayed on its own line?}

\item{name}{(\code{character(1)}) The name of the field.}

\item{name_spec}{(\code{character(1)}, \code{function}, or \code{NULL}) Name specification
passed to \code{\link[vctrs:list_unchop]{vctrs::list_unchop()}}.}

\item{names}{(\code{logical(1)}) Should names be printed even if they can be
deduced from the spec?}

\item{nchar_indent}{(\code{integer(1)}) The number of (additional) characters that
will be used to indent the output when \code{multi_line = TRUE}. Primarily for
internal use when formatting is applied recursively.}

\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}

\item{path_exp}{(\code{list}) The path of the field used as the reference in size
mismatch errors.}

\item{.ptype}{(\code{vector(0)}) A prototype of the desired output type of the
field.}

\item{.ptype_inner}{(\code{vector(0)}) A prototype of the input field.}

\item{.required}{(\code{logical(1)}) Throw an error if the field does not exist?}

\item{simplify_list}{(\code{logical(1)}) Should scalar lists be simplified to
vectors?}

\item{size_act}{(\code{integer(1)}) The observed size of a field.}

\item{size_exp}{(\code{integer(1)}) The expected size of a field.}

\item{spec}{(\code{tspec} or \code{NULL}) A spec object describing the structure of
\code{x}.}

\item{spec_list}{(\code{list}) A list of specifications.}

\item{tib_list}{(\code{list}) A list of tib fields.}

\item{.transform}{(\code{function} or \code{NULL}) A function to apply to the whole
vector after casting to \code{.ptype_inner}.}

\item{value}{(\code{list}) An object list whose fields will be guessed.}

\item{.values_to}{(\code{character(1)} or \code{NULL}) For \code{NULL} (the default), the
field is converted to a \code{.ptype} vector. If a string is provided, the field
is converted to a tibble and the values go into the specified column.}

\item{width}{(\code{integer(1)}) The width (in number of characers) of text output
to generate.}

\item{x}{(\code{any}) The object to check.}
}
\description{
These parameters are used in multiple functions. They are defined here to
make them easier to import and to find.
}
\keyword{internal}
