% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared_params.R
\name{.shared-params}
\alias{.shared-params}
\title{Shared parameters}
\arguments{
\item{.children}{A string giving the name of the field that contains the
children.}

\item{.children_to}{A string giving the column name in which to store the
children.}

\item{.elt_transform}{A function to apply to each element before casting to
\code{.ptype_inner}.}

\item{.fill}{Optionally, a value to use if the field does not exist.}

\item{.format}{Optional, a string passed to the \code{format} argument of
\code{as.Date()}.}

\item{.key}{The path to the field in the object, as a character vector.}

\item{.ptype}{A prototype of the desired output type of the field.}

\item{.ptype_inner}{A prototype of the field.}

\item{.required}{Throw an error if the field does not exist?}

\item{.transform}{A function to apply to the whole vector after casting to
\code{.ptype_inner}.}

\item{.values_to}{Can be one of the following:
\itemize{
\item \code{NULL}: the default. The field is converted to a \code{.ptype} vector.
\item A string: The field is converted to a tibble and the values go into the
specified column.
}}
}
\description{
These parameters are used in multiple functions. They are defined here to
make them easier to import and to find.
}
\keyword{internal}
