% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prep_recursive.R
\name{.spec_prep_recursive}
\alias{.spec_prep_recursive}
\title{Prepare a recursive tibblify specification}
\usage{
.spec_prep_recursive(spec)
}
\arguments{
\item{spec}{(\code{tspec}) A specification of how to convert \code{x}. Generated with
\code{\link[=tspec_df]{tspec_df()}}, \code{\link[=tspec_row]{tspec_row()}}, \code{\link[=tspec_object]{tspec_object()}}, \code{\link[=tspec_recursive]{tspec_recursive()}}, or (if
\code{NULL}, the default), \code{\link[=guess_tspec]{guess_tspec()}}.}
}
\value{
A prepared tibblify specification.
}
\description{
Prepare a recursive tibblify specification
}
\keyword{internal}
