% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_required}
\alias{.stop_required}
\title{Error for missing required field}
\usage{
.stop_required(path)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}
}
\value{
Never returns; called for its side effect of throwing an error.
}
\description{
Error for missing required field
}
\keyword{internal}
