% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_required_colmajor}
\alias{.stop_required_colmajor}
\title{Error for missing required colmajor field}
\usage{
.stop_required_colmajor(path)
}
\arguments{
\item{path}{(\code{list}) A path object encoded as a depth and a list of path
elements.}
}
\value{
\code{NULL} (invisibly). Called for its side effect of throwing an error.
}
\description{
Error for missing required colmajor field
}
\keyword{internal}
