% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.tib_scalar_impl}
\alias{.tib_scalar_impl}
\title{Implementation of tib_scalar}
\usage{
.tib_scalar_impl(
  .key,
  .ptype,
  ...,
  .required = TRUE,
  .fill = vctrs::vec_init(.ptype_inner),
  .ptype_inner = .ptype,
  .transform = NULL,
  .class = NULL,
  .call = caller_env()
)
}
\arguments{
\item{.key}{(\code{character}) The path of names to the field in the object.}

\item{.ptype}{(\code{vector(0)}) A prototype of the desired output type of the
field.}

\item{.required}{(\code{logical(1)}) Throw an error if the field does not exist?}

\item{.fill}{(\code{vector} or \code{NULL}) Optionally, a value to use if the field
does not exist.}

\item{.ptype_inner}{(\code{vector(0)}) A prototype of the input field.}

\item{.transform}{(\code{function} or \code{NULL}) A function to apply to the whole
vector after casting to \code{.ptype_inner}.}

\item{.class}{(\code{character} or \code{NULL}) Additional classes for the collector.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{tib_scalar}) A tibblify scalar collector.
}
\description{
Implementation of tib_scalar
}
\keyword{internal}
