% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec_basics.R
\name{.tib_vector_impl}
\alias{.tib_vector_impl}
\title{Implementation of tib_vector}
\usage{
.tib_vector_impl(
  .key,
  .ptype,
  ...,
  .required = TRUE,
  .fill = NULL,
  .ptype_inner = .ptype,
  .transform = NULL,
  .elt_transform = NULL,
  .input_form = c("vector", "scalar_list", "object"),
  .values_to = NULL,
  .names_to = NULL,
  .class = NULL,
  .call = caller_env()
)
}
\arguments{
\item{.key}{(\code{character}) The path of names to the field in the object.}

\item{.ptype}{(\code{vector(0)}) A prototype of the desired output type of the
field.}

\item{.required}{(\verb{logical(1))} Throw an error if the field does not exist?}

\item{.fill}{(\code{vector} or \code{NULL}) Optionally, a value to use if the field
does not exist.}

\item{.ptype_inner}{(\code{vector(0)}) A prototype of the input field.}

\item{.transform}{(\code{function} or \code{NULL}) A function to apply to the whole
vector after casting to \code{.ptype_inner}.}

\item{.elt_transform}{(\code{function} or \code{NULL}) A function to apply to each
element before casting to \code{.ptype_inner}.}

\item{.input_form}{(\code{character(1)}) The structure of the input field. Can be
one of:
\itemize{
\item \code{"vector"}: The field is a vector, e.g. \code{c(1, 2, 3)}.
\item \code{"scalar_list"}: The field is a list of scalars, e.g. \code{list(1, 2, 3)}.
\item \code{"object"}: The field is a named list of scalars, e.g.
\code{list(a = 1, b = 2, c = 3)}.
}}

\item{.values_to}{(\code{character(1)} or \code{NULL}) For \code{NULL} (the default), the
field is converted to a \code{.ptype} vector. If a string is provided, the field
is converted to a tibble and the values go into the specified column.}

\item{.names_to}{(\code{character(1)} or \code{NULL}) What to do with the inner names
of the object. Can be one of:
\itemize{
\item \code{NULL}: the default. The inner names of the field are not used.
\item A string: Use only if 1) the input form is \code{"object"} or \code{"vector"} and
}
\enumerate{
\item \code{.values_to} is a string. The inner names of the field will populate the
specified column in the field's tibble.
}}

\item{.class}{(\code{character} or \code{NULL}) Additional classes for the collector.}

\item{.call}{(\code{environment}) The environment to use for error messages.}
}
\value{
(\code{tib_vector}) A tibblify vector collector.
}
\description{
Implementation of tib_vector
}
\keyword{internal}
