% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formatting}
\alias{formatting}
\alias{print.tspec}
\alias{format.tspec_df}
\title{Printing tibblify specifications}
\usage{
\method{print}{tspec}(x, width = NULL, ..., names = NULL)

\method{format}{tspec_df}(x, width = NULL, ..., names = NULL)
}
\arguments{
\item{x}{Spec to format or print}

\item{width}{Width of text output to generate.}

\item{...}{These dots are for future extensions and must be empty.}

\item{names}{Should names be printed even if they can be deduced from the
spec?}
}
\value{
\code{x} is returned invisibly.
}
\description{
Printing tibblify specifications
}
\examples{
spec <- tspec_df(
  a = tib_int("a"),
  new_name = tib_chr("b"),
  row = tib_row(
    "row",
    x = tib_int("x")
  )
)
print(spec, names = FALSE)
print(spec, names = TRUE)
}
