% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibblify.R
\name{get_spec}
\alias{get_spec}
\title{Examine the column specification}
\usage{
get_spec(x)
}
\arguments{
\item{x}{The data frame object to extract from.}
}
\value{
A tibblify specification object.
}
\description{
Examine the column specification
}
\examples{
df <- tibblify(list(list(x = 1, y = "a"), list(x = 2)))
get_spec(df)
}
