% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcols.R
\name{lcols}
\alias{lcols}
\title{Create list parser specification}
\usage{
lcols(..., .default = lcol_skip(zap()))
}
\arguments{
\item{...}{Column specification objects created by \verb{lcol_*()}.}

\item{.default}{The default parser to use.}
}
\description{
Create list parser specification
}
\details{
compare to cols, cols_only, col_spec
https://github.com/r-lib/vroom/blob/843e4a66510146120d734980e60cefa550d61ff6/R/col_types.R
}
