% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcols.R
\name{lcols}
\alias{lcols}
\title{Create list parser specification}
\usage{
lcols(..., .default = lcol_skip(zap()))
}
\arguments{
\item{...}{Column specification objects created by \verb{lcol_*()}.}

\item{.default}{The default parser to use.}
}
\description{
Create list parser specification
}
\examples{
lcols(
  lcol_int("id"),
  lcol_chr("name"),
  lcol_lst_of("aliases", character())
)

# To create multiple columns of the same type use the bang-bang-bang (!!!)
# operator together with `purrr::map()`
lcols(
  !!!purrr::map(c("id", "age"), lcol_int),
  !!!purrr::map(c("name", "title"), lcol_chr),
  !!!purrr::map(c("aliases", "addresses"), lcol_guess)
)
}
