% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/should_inform_unspecified.R
\name{should_inform_unspecified}
\alias{should_inform_unspecified}
\title{Determine whether to inform about unspecified fields in spec}
\usage{
should_inform_unspecified()
}
\value{
\code{FALSE} if the option is set to \code{FALSE}, \code{TRUE} otherwise.
}
\description{
Wrapper around \code{getOption("tibblify.show_unspecified")} to return \code{TRUE}
unless the option is explicitly set to \code{FALSE}.
}
