% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspec_guess.R, R/tspec_guess_df.R,
%   R/tspec_guess_object.R
\name{tspec_guess}
\alias{tspec_guess}
\alias{tspec_guess_df}
\alias{tspec_guess_object}
\title{Guess the \code{tibblify()} Specification}
\usage{
tspec_guess(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  inform_unspecified = should_inform_unspecified(),
  call = current_call()
)

tspec_guess_df(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  inform_unspecified = should_inform_unspecified(),
  call = current_call()
)

tspec_guess_object(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  call = current_call()
)
}
\arguments{
\item{x}{A nested list.}

\item{...}{These dots are for future extensions and must be empty.}

\item{empty_list_unspecified}{Treat empty lists as unspecified?}

\item{simplify_list}{Should scalar lists be simplified to vectors?}

\item{inform_unspecified}{Inform about fields whose type could not be
determined?}

\item{call}{The execution environment of a currently running function, e.g.
\code{caller_env()}. The function will be mentioned in error messages as the
source of the error. See the \code{call} argument of \code{\link[=abort]{abort()}} for more
information.}
}
\value{
A specification object that can used in \code{tibblify()}.
}
\description{
Use \code{tspec_guess()} if you don't know the input type.
Use \code{tspec_guess_df()} if the input is a data frame or an object list.
Use \code{tspec_guess_objecte()} is the input is an object.
}
\examples{
tspec_guess(list(x = 1, y = "a"))
tspec_guess(list(list(x = 1), list(x = 2)))

tspec_guess(gh_users)
}
