% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_guess.R, R/spec_guess_df.R,
%   R/spec_guess_list.R, R/spec_guess_object.R, R/spec_guess_object_list.R
\name{spec_guess}
\alias{spec_guess}
\alias{spec_guess_df}
\alias{spec_guess_list}
\alias{spec_guess_object}
\alias{spec_guess_object_list}
\title{Guess the \code{tibblify()} Specification}
\usage{
spec_guess(x, empty_list_unspecified = FALSE, call = current_call())

spec_guess_df(x, empty_list_unspecified = FALSE, call = current_call())

spec_guess_list(
  x,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  call = current_call()
)

spec_guess_object(
  x,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  call = current_call()
)

spec_guess_object_list(
  x,
  empty_list_unspecified = FALSE,
  simplify_list = FALSE,
  call = current_call()
)
}
\arguments{
\item{x}{A nested list.}

\item{empty_list_unspecified}{Treat empty lists as unspecified?}

\item{call}{The execution environment of a currently running function, e.g.
\code{caller_env()}. The function will be mentioned in error messages as the
source of the error. See the \code{call} argument of \code{\link[=abort]{abort()}} for more
information.}

\item{simplify_list}{Try to simplify lists if possible?}
}
\value{
A specification object that can used in \code{tibblify()}.
}
\description{
Guess the \code{tibblify()} Specification
}
\examples{
spec_guess(list(x = 1, y = "a"))
spec_guess(list(list(x = 1), list(x = 2)))

spec_guess(gh_users)
}
