% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_guess.R, R/spec_guess_df.R,
%   R/spec_guess_object.R
\name{spec_guess}
\alias{spec_guess}
\alias{spec_guess_df}
\alias{spec_guess_object}
\title{Guess the \code{tibblify()} Specification}
\usage{
spec_guess(
  x,
  ...,
  empty_list_unspecified = FALSE,
  inform_unspecified = should_inform_unspecified(),
  call = current_call()
)

spec_guess_df(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = TRUE,
  inform_unspecified = should_inform_unspecified(),
  call = current_call()
)

spec_guess_object(
  x,
  ...,
  empty_list_unspecified = FALSE,
  simplify_list = TRUE,
  call = current_call()
)
}
\arguments{
\item{x}{A nested list.}

\item{...}{These dots are for future extensions and must be empty.}

\item{empty_list_unspecified}{Treat empty lists as unspecified?}

\item{inform_unspecified}{Inform about fields whose type could not be
determined?}

\item{call}{The execution environment of a currently running function, e.g.
\code{caller_env()}. The function will be mentioned in error messages as the
source of the error. See the \code{call} argument of \code{\link[=abort]{abort()}} for more
information.}
}
\value{
A specification object that can used in \code{tibblify()}.
}
\description{
Use \code{spec_guess()} if you don't know the input type.
Use \code{spec_guess_df()} if the input is a data frame or an object list.
Use \code{spec_guess_objecte()} is the input is an object.
}
\examples{
spec_guess(list(x = 1, y = "a"))
spec_guess(list(list(x = 1), list(x = 2)))

spec_guess(gh_users)
}
