% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib_spec.R
\name{tib_scalar}
\alias{tib_scalar}
\alias{tib_lgl}
\alias{tib_int}
\alias{tib_dbl}
\alias{tib_chr}
\alias{tib_vector}
\alias{tib_lgl_vec}
\alias{tib_int_vec}
\alias{tib_dbl_vec}
\alias{tib_chr_vec}
\alias{tib_list}
\alias{tib_row}
\alias{tib_df}
\title{Create a Field Specification}
\usage{
tib_scalar(key, ptype, required = TRUE, default = NULL, transform = NULL)

tib_lgl(key, required = TRUE, default = NULL, transform = NULL)

tib_int(key, required = TRUE, default = NULL, transform = NULL)

tib_dbl(key, required = TRUE, default = NULL, transform = NULL)

tib_chr(key, required = TRUE, default = NULL, transform = NULL)

tib_vector(key, ptype, required = TRUE, default = NULL, transform = NULL)

tib_lgl_vec(key, required = TRUE, default = NULL, transform = NULL)

tib_int_vec(key, required = TRUE, default = NULL, transform = NULL)

tib_dbl_vec(key, required = TRUE, default = NULL, transform = NULL)

tib_chr_vec(key, required = TRUE, default = NULL, transform = NULL)

tib_list(key, required = TRUE, default = NULL, transform = NULL)

tib_row(.key, ..., .required = TRUE)

tib_df(.key, ..., .required = TRUE, .names_to = NULL)
}
\arguments{
\item{key, .key}{The path to the field in the object.}

\item{ptype}{A prototype of the desired output type of the field.}

\item{required, .required}{Throw an error if the field does not exist?}

\item{default}{Default value to use if the field does not exist.}

\item{transform}{A function to apply to the field before casting to \code{ptype}.}

\item{...}{Column specification created by \verb{tib_*()} or \verb{spec_*()}.}

\item{.names_to}{A string giving the name of the column which will contain
the names of elements of the object list. If \code{NULL}, the default, no name
column is created}
}
\description{
Use these functions to specify how to convert the fields of an object.
}
\details{
There are basically five different \verb{tib_*()} functions
\itemize{
\item \code{tib_scalar(ptype)}: Cast the field to a length one vector of type \code{ptype}.
\item \code{tib_vector(ptype)}: Cast the field to an arbitrary length vector of type \code{ptype}.
\item \code{tib_list()}: Cast the field to a list.
\item \code{tib_row()}: Cast the field to a named list.
\item \code{tib_df()}: Cast the field to a tibble.
}

There are some special shortcuts of \code{tib_scalar()} resp. \code{tib_vector()} for
the most common prototypes
\itemize{
\item \code{logical()}: \code{tib_lgl()} resp. \code{tib_lgl_vec()}
\item \code{integer()}: \code{tib_int()} resp. \code{tib_int_vec()}
\item \code{double()}: \code{tib_dbl()} resp. \code{tib_dbl_vec()}
\item \code{character()}: \code{tib_chr()} resp. \code{tib_chr_vec()}
}
}
\examples{
tib_int("int")
tib_int("int", FALSE, default = 0)

tib_scalar("date", Sys.Date(), transform = function(x) as.Date(x, format = "\%Y-\%m-\%d"))

tib_df(
  "data",
  .names_to = "id",
  age = tib_int("age"),
  name = tib_chr("name")
)
}
