% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibblify-package.R
\docType{package}
\name{tibblify-package}
\alias{tibblify}
\alias{tibblify-package}
\title{tibblify: Rectangle Nested Lists}
\description{
Tool to rectangle nested lists automatically or according to a given specification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mgirlich/tibblify}
  \item Report bugs at \url{https://github.com/mgirlich/tibblify/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Girlich \email{maximilian.girlich@metoda.com}

}
\keyword{internal}
