% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibblify.R
\name{tibblify}
\alias{tibblify}
\title{Rectangle a nested list}
\usage{
tibblify(x, spec = NULL, names_to = NULL)
}
\arguments{
\item{x}{A nested list.}

\item{spec}{A specification how to convert \code{x}. Generated with \code{spec_row()}
or \code{spec_df()}.}
}
\value{
Either a tibble or a list, depending on the specification
}
\description{
Rectangle a nested list
}
\examples{
x <- list(
  list(id = 1, name = "Tyrion Lannister"),
  list(id = 2, name = "Victarion Greyjoy")
)

tibblify(x)
}
