% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untibblify.R
\name{untibblify}
\alias{untibblify}
\title{Convert a data frame or object into a nested list}
\usage{
untibblify(x, spec = NULL)
}
\arguments{
\item{x}{A data frame or an object.}

\item{spec}{Optional. A spec object which was used to create \code{x}.}
}
\value{
A nested list.
}
\description{
The inverse operation to \code{tibblify()}. It converts a data frame or an object
into a nested list.
}
\examples{
x <- tibble(
  a = 1:2,
  b = tibble(
    x = c("a", "b"),
    y = c(1.5, 2.5)
  )
)
untibblify(x)
}
